% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{add_dictionary_meta_data}
\alias{add_dictionary_meta_data}
\title{Add Dictionary Meta-data to Column Attributes}
\usage{
add_dictionary_meta_data(
  dtf,
  column_name,
  description = "",
  variable_category = NULL,
  data_type = NULL,
  sub_category = "",
  redcap_variables = "",
  values_and_labels = "",
  scale_details = "",
  subscale_details = "",
  summary_of_x = "",
  codes_for_missing = list(NA, "", " ", as.Date("1970-01-01", format = "\%Y-\%m-\%d")),
  units_of_x = "",
  validated = "",
  notes = "",
  group_var = NULL,
  time_var = NULL,
  study_var = NULL,
  abbr_labels = list(),
  known_scale = NULL,
  digits = 2
)
}
\arguments{
\item{dtf}{A data frame.}

\item{column_name}{A character string, the name of the
column to add attributes to.}

\item{description}{A character string, a human-readable
description of the variable.}

\item{variable_category}{The overarching category to which
a variable belongs (e.g., session details, subject details,
etc.). By default, the function infers the category using
the initial 3-letter abbreviation contained in the column
name.}

\item{data_type}{The type of data in a column (e.g.,
integer, logical, character string, etc.). By default, the
function infers the data type from the second 3-letter
abbreviation contained in the column name.}

\item{sub_category}{The label for an optional sub-category
of abbreviation (e.g., time points for wide-form data).}

\item{redcap_variables}{A character string, the variable
name in the raw REDCap data from which the current
variable was derived.}

\item{values_and_labels}{An optional named list of two vectors:
\enumerate{
  \item Values: A vector of the unique values (e.g., numeric
  codes, abbreviations, logical values) stored in the column;
  \item Labels: A character vector with the descriptive labels
  for each corresponding value.
}}

\item{scale_details}{An optional list with information for
scale/inventory/questionnaire scores. See the functions
\code{\link{known_scales}} and \code{\link{scale_format}}
on standardized formats for the input.}

\item{subscale_details}{An optional list with information for
subscale scores. See the functions
\code{\link{known_scales}} and \code{\link{scale_format}}
on standardized formats for the input.}

\item{summary_of_x}{A character string, either...
\itemize{
  \item \code{continuous}: The mean, standard
    deviation, min, 16% quantile, 1st quartile,
    median, 3rd quartile, 84% quantile, and max,
    along with the frequency and percent missing;
  \item \code{categorical}: The frequency and percent
    for each category, along with frequency and
    percent missing.
  \item \code{range}: The min and max values, along
    with frequency and percent missing.
}}

\item{codes_for_missing}{A list with the codes used to
indicate missing values (e.g., \code{NA} or \code{''}).}

\item{units_of_x}{A character string detailing the units
of measurement for the variable (e.g., height in cm,
summed scores, likert-scale response, etc.).}

\item{validated}{A character string, notes on whether the
variable has been data-checked.}

\item{notes}{An character string, a free text field for
any additional notes regarding the variable.}

\item{group_var}{An optional character string giving
the column in \code{dtf} containing the group labels
(e.g., labels for intervention versus control groups).}

\item{time_var}{An optional character string giving the
the column in \code{dtf} containing the different
time points (e.g., baseline versus follow-up time points).}

\item{study_var}{An optional character string giving
the column in \code{dtf} containing the study labels
(e.g., labels for pilot versus active studies).}

\item{abbr_labels}{An optional named list (either
\code{variable_category}, \code{date_type}, or
\code{sub_category}) containing the matrix to
pass to the \code{custom} argument for the
\code{\link{column_abbreviations}} function.}

\item{known_scale}{An optional argument to pass
along output from the \code{\link{known_scales}}
function, which can be used in lieu of the
\code{scale_details} and \code{subscale_details}
arguments.}

\item{digits}{Number of digits to round to for summaries.}
}
\value{
An updated data frame where the specified column
has with a new attribute of class 'dictionary_meta_data'
with the data dictionary information.
}
\description{
Updates the attributes for a column in a data frame
with a standardized list of details on the variable
that can be then used to create a data dictionary.
Columns are assumed to be named according to a
standardized scheme, typically in the style
'GGG.TTT.Var_name' or 'GGG.TTT.Optional.Var_name',
where 'GGG' is a 3-letter abbreviation giving the
overarching group for a variable
(e.g., 'Identifiers', 'Session details', etc.),
'TTT' is a 3-letter abbreviation for the type of data
(e.g., 'Integer', 'Character string', etc.),
'Optional' is an optional sub-category (e.g.,
indicators for time points like 'BL' for baseline
and 'Y1' or 'Y2' or subsequent years), and
'Var_name' is a human-readable label
(e.g., 'Treatment_groups', 'Session_dates', etc.).
}
\author{
Kevin Potter
}
