% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{add_meta_data}
\alias{add_meta_data}
\title{Add Meta-data to a Column for Data Dictionary Purposes}
\usage{
add_meta_data(
  dtf,
  column_name,
  description = "",
  variable_category = NULL,
  data_type = NULL,
  values = "",
  labels_for_values = "",
  scales = NULL,
  subscales = NULL,
  range_of_x = NULL,
  summary_of_x = NULL,
  category_labels = NULL,
  type_labels = NULL,
  codes_for_missing = c("", " "),
  group_var = NULL,
  time_points_var = NULL,
  additional_abbr = NULL,
  units_of_x = "",
  digits = 2
)
}
\arguments{
\item{dtf}{A data frame.}

\item{column_name}{...}

\item{description}{...}

\item{variable_category}{...}

\item{data_type}{...}

\item{values}{...}

\item{labels_for_values}{...}

\item{scales}{...}

\item{subscales}{...}

\item{range_of_x}{...}

\item{summary_of_x}{...}

\item{category_labels}{...}

\item{type_labels}{...}

\item{codes_for_missing}{...}

\item{group_var}{...}

\item{time_points_var}{...}

\item{additional_abbr}{...}

\item{units_of_x}{...}

\item{digits}{...}
}
\value{
An updated data frame with new attributes for
the specified column.
}
\description{
Updates the attributes for a column in a data frame
with a standardized list of details on the variable
that can be then used to create a data dictionary.
Columns are assumed to be named according to a
standardized scheme, typically in the style
'AAA.BBB.Description', where 'AAA' is a 3-letter
abbreviation giving the overarching category for
a variable (e.g., 'Identifiers', 'Session details', etc.),
'BBB' is a 3-letter abbreviation for the type of data
(e.g., 'Integer', 'Character string', etc.), and
'Description' is a human-readable label
(e.g., 'Treatment_groups', 'Session_dates', etc.).
}
\author{
Kevin Potter
}
