% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{Assignment}
\alias{Assignment}
\alias{\%+=\%}
\alias{\%-=\%}
\title{Addition/Subtraction Assignment Operators}
\usage{
x \%+=\% y

x \%-=\% y
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
A numeric vector.
}
\description{
The operator \code{\%+=\%} adds the right operand
to a variable and assigns the resulting value to
the variable. The operator \code{\%-=\%} subtracts
the right operand from a variable and assigns the
resulting value to the variable.
}
\details{
Care must be taken with order of operations.
Because \code{\%+=\%} and \code{\%-=\%} are functions,
they and their arguments will be evaluated first,
followed by subsequent operations. Hence a call like
\code{ x \%+=\% 1/2 } will not result in an increment of
0.5 to x. Instead, x will be first be
incremented by 1. Then the call \code{ x/2 } will be
run with no assignment of values.
}
\examples{
# Simple assignment
x <- 1
x \%+=\% 1
x
y <- 1
y \%-=\% 2
y

# Order of operations can be tricky
x <- 1
y <- 1
invisible(x \%+=\% y / 2)
x
# Above is equivalent to (x \%+=\% y)/2

# Therefore embed multiple operations in parentheses
x <- 1
y <- 1
x \%+=\% (y / 2)
x

# Vectorized
x <- 1:3
x \%+=\% 3
x
x <- 3:1
x \%-=\% 2:0
x
}
\references{
https://stackoverflow.com/questions/5738831
}
\author{
Kevin Potter
}
