% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{bin_age}
\alias{bin_age}
\title{Helper function to bin ages (used when processing sample and census data)}
\usage{
bin_age(x, breaks, labels = NULL, right = TRUE)
}
\description{
Helper function to bin ages (used when processing sample and census data)
}
