% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_add_attribute}
\alias{camr_add_attribute}
\title{Add Attribute to R Object}
\usage{
camr_add_attribute(x, attribute, label)
}
\arguments{
\item{x}{An R object.}

\item{attribute}{The element to add
as an attribute.}

\item{label}{The label to use in the
list of attributes.}
}
\value{
The R object with updated attributes.
}
\description{
Function to add an attribute to an
R object while preserving any
existing attributes (especially
useful for not overwriting attributes
for dates to function properly).
}
\examples{

x <- 1:3
x <- x |> camr_add_attribute( "Hello world", "example" )
print( x )

# Preserves existing attributes
x <- as.Date( c( "2000-01-01", "2010-06-15", "2020-12-25" ) )
x <- x |> camr_add_attribute( "Hello world", "example" )
print( class( x ) )
print( attributes( x )$example )

}
\author{
Kevin Potter
}
