% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_add_codebook_entry}
\alias{camr_add_codebook_entry}
\alias{camr_ace}
\title{Add a Codebook Entry to a Data Frame}
\usage{
camr_add_codebook_entry(
  dtf,
  variable,
  description = "",
  category = "",
  data_type = "",
  subcategory = "",
  values_and_labels = "",
  inventory = "",
  units_of_measurement = "",
  descriptive_summary = "",
  codes_for_missing_data = NULL,
  deidentified = "FALSE",
  collected_over = "",
  created_from = "",
  data_quality_checks = "",
  notes = "",
  custom_category = NULL,
  custom_data_type = NULL,
  custom_subcategory = NULL,
  missing_attribute_label = "missing",
  digits = 2,
  add_missing_to_attr = NULL,
  non_standard = TRUE
)

camr_ace(...)
}
\arguments{
\item{dtf}{A data frame, preferably one using
standardized naming conventions for columns
(see \link{camr_column_abbreviations}).}

\item{variable}{The column name for the variable
whose attributes should be updated to include
a codebook entry. Non-standard evaluation possible.}

\item{description}{A character string, a brief human-readable
description of what the variable is.}

\item{category}{An optional character string, the
category label for the variable. If left blank, the
function automatically determines the category via
\link{camr_column_abbreviations}.}

\item{data_type}{An optional character string, the
data type label for the variable. If left blank, the
function automatically determines the data type via
\link{camr_column_abbreviations}.}

\item{subcategory}{An optional character string, the
subcategory label for the variable. If left blank, the
function automatically determines the subcategory via
\link{camr_column_abbreviations}.}

\item{values_and_labels}{Either the empty character set or
a named list of two vectors, (1) \code{content} containing
the distinct values for the variable, and (2)
\code{additional_content} containing the label associated
with each value in (1).}

\item{inventory}{Either (1) output from the
\link{camr_inventories} function, (2) the empty character
set, or (3) a named list of two vectors, where \code{content}
contains details on inventory an \code{additional_content}
labels each type of detail (name, number of items, etc.).}

\item{units_of_measurement}{A character string, the
units of measurement for the variable.}

\item{descriptive_summary}{Either the empty character set or
a named list of two vectors, (1) \code{content} containing
summary statistics for the variable, and (2)
\code{additional_content} containing the label associated
with each statistic in (1).}

\item{codes_for_missing_data}{An optional vector with
codes for missing data. If \code{NULL} uses default
codes. If \code{variable} has an attribute with the
codes, the codes in the attribute list override
this argument.}

\item{deidentified}{Either \code{FALSE} if the variable
is not deidentified (i.e., contains potential patient
health information) or \code{TRUE} otherwise.}

\item{collected_over}{Either the empty character set or
a named list of two vectors, (1) \code{content} containing
distinct values for groups, time points, etc., over which
the variable was collected, and (2) \code{additional_content}
containing the column name associated with the values in (1).}

\item{created_from}{An optional character vector, the original
REDCap variables from which the variable was created.}

\item{data_quality_checks}{An optional character string, notes
on any data quality checks relevant to the variable.}

\item{notes}{An optional character vector, any additional notes
to include in the codebook entry.}

\item{custom_category}{An optional matrix with two
columns, 'Abbr' and 'Label', for custom abbreviations
and labels for variable categories.}

\item{custom_data_type}{An optional matrix with two
columns, 'Abbr' and 'Label', for custom abbreviations
and labels for data types.}

\item{custom_subcategory}{An optional matrix with two
columns, 'Abbr' and 'Label', for custom abbreviations
and labels for variable subcategories.}

\item{missing_attribute_label}{A character string, the
name of the element in the list of attributes
containing codes for missing data.}

\item{digits}{The number of decimal places to round to
for the descriptive summaries.}

\item{add_missing_to_attr}{An optional logical value;
if \code{FALSE} will not add codes for missing data
to the attributes of \code{Variable}. If \code{NULL}
defaults to adding if no pre-existing attribute is found.}

\item{non_standard}{A logical value; if \code{TRUE}
uses non-standard evaluation for \code{variable}.}
}
\value{
The data frame \code{dtf}, with the attributes for
the column specified by \code{variable} updated to
contain a data frame of additional class
\code{codebook_entry}.
}
\description{
Function to update the attributes for a column
in a data frame to contain a codebook entry.
This ensures easy portability (at least within the
framework of R) of the codebook for a data set.
}
\examples{
data( "example_CAM_data_set" )
dtf <- example_CAM_data_set

# Add codebook entry to data frame
dtf <- dtf |>
  camr_add_codebook_entry(
    IDS.INT.Screening,
    description = "Identifier assigned at screening session"
  )

dtf |> camr_pull_codebook_entry( IDS.INT.Screening )

dtf <- dtf |>
  camr_add_codebook_entry(
    IDS.CHR.Participant,
    "Identifier assigned at enrollment"
  )

dtf |> camr_pull_codebook_entry( IDS.CHR.Participant )

dtf <- dtf |>
  camr_add_codebook_entry(
    INV.INT.HADS_anxiety,
    inventory = camr_inventories( "HADS", "Anxiety" ),
    descriptive_summary = "continuous",
    collected_over = camr_ice(
      "combine",
      dtf |> camr_collected_over( INV.INT.HADS_anxiety, SSS.CHR.Session ),
      dtf |> camr_collected_over( INV.INT.HADS_anxiety, SSS.CHR.Event )
    )
  )
dtf |> camr_pull_codebook_entry( INV.INT.HADS_anxiety )

}
\author{
Kevin Potter
}
