% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R06-Inventories_scales_questionnaires.R
\name{camr_add_inventory_item}
\alias{camr_add_inventory_item}
\title{Add Individual Inventory Item With Codebook Entry}
\usage{
camr_add_inventory_item(
  dtf_original,
  dtf_new,
  chr_original_column,
  chr_new_column,
  vec_response_options = NULL,
  vec_new_values = NULL,
  lst_values_and_labels = "",
  chr_inventory = "",
  chr_question = "",
  chr_description = "",
  ...
)
}
\arguments{
\item{dtf_original}{A data frame with the
column of scores for the given item or question.}

\item{dtf_new}{A data frame to which the column
of scores will be added.}

\item{chr_original_column}{A character string, the
column name in \code{dtf_original} to add.}

\item{chr_new_column}{A character string, the
new column name to use in \code{dtf_new}.}

\item{vec_response_options}{A vector, the original
response options. Used to recode the response
options if needed.}

\item{vec_new_values}{A vector, the new values to
replace \code{vec_response options} with. Must
match in length to \code{vec_response_options}.}

\item{lst_values_and_labels}{A named list with
\code{'content'} containing the numeric values
and \code{'additional_content'} containing the
labels or interpretations for the values.}

\item{chr_inventory}{A character string, the
name of the inventory, survey, or questionnaire
to which the item belongs.}

\item{chr_question}{A character string, the item's
framing question.}

\item{chr_description}{A character string,
allows a user to specify a custom description
rather than one auto-filled using \code{chr_inventory}
and \code{chr_question}.}

\item{...}{Additional arguments for the
\link{camr_add_codebook_entry} function.}
}
\value{
A data frame with an additional column
containing the scores for the individual item
or question.
}
\description{
Convenience function to add a column with scores
for an individual item or question from an
inventory, survey, or questionnaire. Can also
add the associated codebook entry for the item.
}
