% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{camr_cache}
\alias{camr_cache}
\title{Create a global cache object.}
\usage{
camr_cache()
}
\description{
Creates a global object \code{camr} for storing information such as file paths in
an encapsulated manner. If a \code{.cache} Rds file can be found in the working
directory, its contents will be loaded.
}
\author{
Michael Pascale
}
