% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R22-yaml_codebook_maker.R
\name{camr_check_codebook}
\alias{camr_check_codebook}
\title{Check codebook for consistency with data}
\usage{
camr_check_codebook(data_df, codebook_df)
}
\arguments{
\item{data_df}{A dataframe of data that should match the codebook}

\item{codebook_df}{A dataframe containing the codebook}
}
\value{
(Invisibly) Logical vector. Indicates whether each variable in
\code{data_df} contains unexpected values (based on the codebook).
}
\description{
Check that dataframe data matches expectations established by codebook.
Intended for use during initial generation of codebooks and also after
later updates of codebook. Prints warnings if matching expectations are not met.
}
