% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_check_name_conventions}
\alias{camr_check_name_conventions}
\title{Check Naming Conventions}
\usage{
camr_check_name_conventions(df_any)
}
\arguments{
\item{df_any}{A dataframe.}
}
\value{
Prints invalid column names to the console.
}
\description{
Checks naming conventions of a processed dataframe,
following standards set on the CAM Wiki.
}
\author{
Michael Pascale
}
