% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04-REDCap.R
\name{camr_ckdict}
\alias{camr_ckdict}
\title{Verify REDCap data dictionary.}
\usage{
camr_ckdict(dict)
}
\arguments{
\item{dict}{REDCap metadata table as downloaded by [redcap_read()].}
}
\description{
Verify REDCap data dictionary.
}
\examples{

}
\author{
Michael Pascale
}
