% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_collected_over}
\alias{camr_collected_over}
\title{Categories over which Data was Collected}
\usage{
camr_collected_over(dtf, x, y, missing_values = NULL, non_standard = TRUE)
}
\arguments{
\item{dtf}{A data frame.}

\item{x}{The variable for the collected data (non-standard
evaluation possible).}

\item{y}{The variable with the categories (non-standard
evaluation possible).}

\item{missing_values}{An optional logical vector indicating
missing data (must match in length to \code{x}). If
\code{NULL} assumes \code{NA} values are the only
indicator of missing data.}

\item{non_standard}{A logical value; if \code{TRUE}
uses non-standard evaluation for \code{x} and \code{y},
otherwise assumes they are character strings.}
}
\value{
A list with two vectors: \code{content} with the
categories from \code{y} over which there was non-missing
data, and \code{additional_content} listing the column
name for \code{y}.
}
\description{
Function to list the categories (e.g., experimental
groups, study time points) over which data was collected.
Useful for denoting the specific subset of categories data
was collected over (e.g., if data was only collected at a
baseline time point). Output can be passed to the
\code{Collected_over} argument for the
\link{camr_new_codebook_entry} function.
}
\examples{
data( example_CAM_data_set )

example_CAM_data_set |>
  camr_collected_over( INV.INT.HADS_anxiety, SSS.CHR.Session )

example_CAM_data_set |>
  camr_collected_over( INV.DBL.BPI_severity, SSS.CHR.Event )

}
\author{
Kevin Potter
}
