% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_combine_redcap_checkboxes}
\alias{camr_combine_redcap_checkboxes}
\title{Combine REDCap Checkbox Fields into one Column}
\usage{
camr_combine_redcap_checkboxes(
  cols,
  fn_rename = function(x) str_remove(x, "^.*___"),
  fn_predicate = function(x) {
     if (!all(x \%in\% 0:1)) 
        
    stop("Specified columns are not all 0 or 1.")
     x == "1"
 },
  collapse = "|",
  as.factor = FALSE
)
}
\arguments{
\item{cols}{Tidyselect expression.}

\item{fn_rename}{Function with which to rename collapsed columns. Default removes the '___' prefix.}

\item{fn_predicate}{Function which transforms each checkbox field value to TRUE/FALSE. Default is x == 1.}

\item{collapse}{Character to separate collapsed values. Defaut is '|'.}

\item{as.factor}{Create a factor where the levels are all the possible combinations of values. Default is FALSE.}
}
\value{
Character or factor vector of collapsed checkbox values.
}
\description{
Use this function within \code{dplyr::mutate}.
}
