% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_commit_number}
\alias{camr_commit_number}
\title{Determine Last Commit Number for Current Branch}
\usage{
camr_commit_number()
}
\value{
A character string, the alphanumeric
code for the current commit.
}
\description{
Function to extract the 7-digit alphanumeric code
for the most recent commit associated with the
current branch.
}
