% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S00-Globals.R
\docType{data}
\name{camr_const_datetime_regex}
\alias{camr_const_datetime_regex}
\title{Regular expression to extract datetimes from strings.}
\format{
An object of class \code{character} of length 1.
}
\usage{
camr_const_datetime_regex
}
\description{
Expects dates dash or forward-slash delimited and optionally followed by a
time component. Use with \code{\link[stringr:str_extract]{stringr::str_extract()}} to pull out datetimes from
character vectors.
}
\author{
Michael Pascale
}
\keyword{datasets}
