% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_copy_from_source}
\alias{camr_copy_from_source}
\title{Copy Files From Source Folder}
\usage{
camr_copy_from_source(
  chr_path_to_source_folder = "",
  chr_environment = "FOLDER_SOURCE",
  chr_match_subfolder = "",
  chr_new_folder = "Source"
)
}
\arguments{
\item{chr_environment}{An optional character
string, the environmental variable with
the path to the source folder.}

\item{chr_match_subfolder}{An optional character
string, a pattern to match against for selecting
a desired subfolder. Otherwise, function takes
most recent subfolder.}

\item{chr_new_folder}{A character string, the
name of the folder in which to store the
copied files for the current directory.}

\item{chr_path_to_folder}{A character string,
the absolute path to the source folder.}
}
\value{
As a side effect copies files to a
new subfolder in the current directory.
}
\description{
Function to copy files from a subfolder
in a source folder to a new subfolder in
a user-defined source folder in the current
directory.
}
