% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{camr_data_checking_report}
\alias{camr_data_checking_report}
\title{Process REDCap logging data. Produce plots and tables summarizing data checking procedures.}
\usage{
camr_data_checking_report(
  df_sl_subject_ids,
  df_vl_visits,
  df_redcap_raw,
  chr_rc_uri = "",
  chr_rc_token = "",
  record_col = IDS.CHR.Record,
  visit_date_col = SSS.DAT.Visit,
  users_exclude = NULL,
  out = NULL
)
}
\arguments{
\item{df_sl_subject_ids}{A dataframe from the data processing pipeline containing record and subject identifiers.}

\item{df_vl_visits}{A dataframe from the data processing pipeline containing visit dates.}

\item{df_redcap_raw}{A dataframe containing raw REDCap study data outputted by the data processing pipeline.}

\item{chr_rc_uri}{A character string, the URL to use for the httr::POST function. Pulls from the standard \code{.renviron} file by default.}

\item{chr_rc_token}{A character string, the user's API token for the specified REDCap project. Pulls from the standard \code{.renviron} file by default.}

\item{record_col}{A column name specifying the record IDs in df_sl_subject_ids.}

\item{visit_date_col}{A column name specifying the visit dates in df_vl_visits.}

\item{users_exclude}{An optional vector of REDCap usernames listing who to exclude from data quality rule execution.}

\item{out}{An optional file path to save out an HTML file and two SVG files containing the table and plots.}
}
\value{
A list of objects containing two ggplot objects visualizing data checking times and one gt table presenting the last three dates that data quality rules were executed.
}
\description{
Function that downloads and processes REDCap logging data and joins with data processing pipeline standard outputs to produce figures reporting progress on a study's data checking procedures. Figures include a table presenting the last 3 dates that data quality rules, a histogram depicting the amount of time since a data query was opened, and a histogram depicting the amount of time between a study visit and manual data checking for that visit. By default, the function will output a list of objects containing these figures, and you can set a file path \code{out} to save out these figures as well.
}
\examples{
# at minimum, the following three dataframes from a data processing pipeline in the standard format are required
lst_data_checking_figs <- camr_data_checking_report(df_sl_subject_ids, df_vl_visits, df_redcap_raw)

# additional arguments can be specified to exclude users' actions (i.e., programmers testing data quality rules) from reports or to output files in a specified location
lst_data_checking_figs <- camr_data_checking_report(df_sl_subject_ids, df_vl_visits, df_redcap_raw, users_exclude = c('be931', 'll278', 'zh121'), out = file.path(getwd(), 'data/output'))

# each figure can be extracted from the list to include in a progress report
lst_data_checking_figs$gt_dqr_dates
lst_data_checking_figs$p_query_hist
lst_data_checking_figs$p_manual_hist

}
\author{
Bryn Evohr
}
