% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_data_frame_from_codebook_entry}
\alias{camr_data_frame_from_codebook_entry}
\title{Create Data Frame From Codebook Entries}
\usage{
camr_data_frame_from_codebook_entry(dtf)
}
\arguments{
\item{dtf}{A data frame whose variables have
codebook entries.}
}
\value{
A data frame with all codebook entries.
}
\description{
Function that creates a data frame containing
all codebook entries - useful for creating
a .csv codebook.
}
\author{
Kevin Potter
}
