% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_data_frame_to_codebook_entry}
\alias{camr_data_frame_to_codebook_entry}
\title{Convert Data Frame to Codebook Entry}
\usage{
camr_data_frame_to_codebook_entry(dtf_main, dtf_codebook)
}
\arguments{
\item{dtf_main}{A data frame of observations.}

\item{dtf_codebook}{A data frame with codebook
entries (see \link{camr_data_frame_from_codebook_entry}).}
}
\value{
A data frame with codebook entries as
attributes for the relevant variables.
}
\description{
Function to extract codebook entries from a
data frame and add them to the attributes for
the relevant variables in a data frame of
observations.
}
\author{
Kevin Potter
}
