% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_deidentified_codebook_entry}
\alias{camr_deidentified_codebook_entry}
\title{Flag De-identified Variables}
\usage{
camr_deidentified_codebook_entry(dtf)
}
\arguments{
\item{dtf}{A data frame.}
}
\value{
A logical vector matching in length to the number
of colunns in \code{dtf}, equal to \code{TRUE} for
any variables with codebook entries indicating the variable
is de-identified.
}
\description{
Function to flag whether a variable in a data frame
is de-identified (i.e., does not risk exposure of
patient health information) based on codebook entries.
}
\details{
According to the Health Insurance Portability
and Accountability Act of 1996 (HIPAA), under the
'Safe Harbor' method of de-identification, there are
18 types of identifiers that should be removed for
a data set to avoid risk of exposing patient health
information (PHI):
\itemize{
\item Names
\item All geographic subdivisions smaller than a state
\item All elements of dates (except year) directly
related to an individual
\item Telephone numbers
\item Vehicle identifiers and serial numbers
\item Fax numbers
\item Device identifiers and serial numbers
\item Email addresses
\item Web Universal Resource Locators (URLs)
\item Social security numbers
\item Internet Protocol (IP) addresses
\item Medical record numbers
\item Biometric identifiers
\item Health plan beneficiary numbers
\item Full-face photographs and comparable images
\item Account numbers
\item Any other unique identifying number, characteristic, or code
\item Certificate/license numbers
}
}
\examples{
data( "example_CAM_data_set" )

}
