% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_descriptive_summary}
\alias{camr_descriptive_summary}
\title{Create Descriptive Summary of Data}
\usage{
camr_descriptive_summary(x, type, missing_values = NULL, digits = 2)
}
\arguments{
\item{x}{A variable.}

\item{type}{The type of summary to return,
either \code{continuous} (mean, SD, etc.),
\code{categorical} (percentages, etc.), or
\code{range} (date ranges).}

\item{missing_values}{An optional logical
vector matching in length to \code{x}
indicating missing data.}

\item{digits}{The number of digits to round to.}
}
\value{
A list with two vectors: \code{content}
with the summary statistics and \code{additional_content}
with the labels for each type of statistic.
}
\description{
Function to generate descriptive summaries
(means, percentages, ranges, etc.) of a
variable that can then be passed to the
\code{descriptive_summary} argument of the
\link{new_codebook_entry} function.
}
\examples{
data( example_CAM_data_set )

example_CAM_data_set$INV.INT.HADS_anxiety |>
  camr_descriptive_summary( "continuous") |>
  data.frame()

example_CAM_data_set$SBJ.INT.Biological_sex |>
  camr_descriptive_summary( "categorical") |>
  data.frame()

}
\author{
Kevin Potter
}
