% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_display_codebook_entry}
\alias{camr_display_codebook_entry}
\alias{camr_dice}
\title{Display Codebook Entry}
\usage{
camr_display_codebook_entry(x, variable = "")

camr_dice(...)
}
\arguments{
\item{x}{Either (a) a data frame or (b) a variable with
a \code{codebook_entry} attribute.}

\item{variable}{If \code{x} is a data frame, the column
in the data frame from which the codebook entry
should be extracted and displayed.}
}
\value{
Output to the console window detailing the
codebook entry.
}
\description{
Print details from a codebook entry to the console window.
}
\examples{
data( "example_CAM_data_set" )
dtf <- example_CAM_data_set[,1:2]

# Add codebook entry to data frame
dtf <- dtf |>
  camr_add_codebook_entry(
    IDS.INT.Screening,
    description = "Identifier assigned at screening session"
)

# Display entry
dtf |> camr_display_codebook_entry( IDS.INT.Screening )

# Alternative approach
camr_display_codebook_entry( dtf$IDS.INT.Screening )

}
\author{
Kevin Potter
}
