% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_file_name}
\alias{camr_file_name}
\title{Generate a Standardized File Name}
\usage{
camr_file_name(
  description,
  extension = NULL,
  project = NULL,
  date = NULL,
  time = NULL,
  datetime = NULL,
  git = TRUE
)
}
\arguments{
\item{description}{A character string giving
a brief description or human-readable name
for the file (preferably in the format
\code{File_name}).}

\item{extension}{An optional character string
giving the file extension sans period. If
\code{NULL} defaults to \code{RData}.}

\item{project}{An optional character string with
the project label. If \code{NULL} will attempt to
extract from an environment variable or a config file.}

\item{date}{An optional character string giving
the date (typically in the format YYYY_MM_DD).}

\item{time}{An optional character string giving
the time (typically in the format HH_MM_SS).}

\item{datetime}{An optional character string giving
both the date and time (typically in the format
YYYY_MM_DD-HH_MM_SS).}

\item{git}{Logical; if \code{TRUE} includes the
most recent commit number.}
}
\value{
A character string, a file name of the form
\code{project-description-date-time-commit.extension}.
}
\description{
Function to generate a file name of the form
\code{project-description-date-time-commit.extension}.
}
\author{
Michael Pascale, Kevin Potter
}
