% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_file_path}
\alias{camr_file_path}
\title{Determine Absolute Path to a File}
\usage{
camr_file_path(partial_name, env_variable = NULL, folder = NULL, latest = TRUE)
}
\arguments{
\item{partial_name}{A character string, the file name
or part of the file name that can uniquely identify
it from other files.}

\item{env_variable}{An optional character string, an
environmental variable listed in a .env file
loaded into the R environment via the \strong{dotenv}
package. If provided, will be used as the source
folder assumed to contain the file.}

\item{folder}{An optional character string, the
name of the folder (or part of the folder name
that can uniquely identify it) containing the
file. If the \code{env_variable} argument is provided,
then the \code{folder} argument can be used to specify
a subfolder containing the file.}

\item{latest}{Logical; if \code{TRUE} the function will
sort all matching file names and return the final
one (if file names contain dates, this means
the function returns the most recent file).}
}
\value{
A character string, an absolute path to a file.
}
\description{
Function to find the absolute path to a file given a
partial match to its name.
}
