% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08-Templates.R
\name{camr_file_templates}
\alias{camr_file_templates}
\title{Create File Template}
\usage{
camr_file_templates(chr_template, chr_file_name = "")
}
\arguments{
\item{chr_template}{A character string specifying
the type of template to create, either...
\itemize{
\item renviron = A .renviron file.
\item targets = A _targets.R file.
}}

\item{chr_file_name}{An optional character
string, the name for the new file.}
}
\value{
As a side effect, a new file.
}
\description{
Function to create a template
for various type of files.
}
