% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S06-FilesAndPaths.R
\name{camr_filename}
\alias{camr_filename}
\title{Generates a standard filename.}
\usage{
camr_filename(
  description = stop("Must provide file description."),
  extension = stop("Must provide file extension."),
  project = camr_get("project"),
  date = format(Sys.time(), "\%Y\%m\%d", tz = "UTC"),
  time = format(Sys.time(), "\%H\%M\%S", tz = "UTC"),
  git = TRUE
)
}
\arguments{
\item{description}{String. Preferably UpperCamelCase with no spaces.}

\item{extension}{String. The file extension(s) with no leading \code{.}.}

\item{project}{Optional string. Name of the project. Defaults to that stored
in the global cache object.}

\item{date}{Optional string. Date in YYYYMMDD format. Defaults to the current
date.}

\item{time}{Optional string. Time in HHMMSS format. Defaults to the current
time in UTC.}

\item{git}{Optional logical. The project is a git repository. Defaults to T.}
}
\value{
String. The generated filename.
}
\description{
Generates a standard filename.
}
\author{
Michael Pascale
}
