% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R14-Yaml_codebook.R
\name{camr_fill_codebook_from_redcap}
\alias{camr_fill_codebook_from_redcap}
\title{Use variable mappings from camr_map_redcap_vars to fill in codebook info}
\usage{
camr_fill_codebook_from_redcap(
  api_token_path,
  codebook_dir,
  mapping_file,
  output_dir
)
}
\arguments{
\item{api_token_path}{path to text file containing REDCap project API token}

\item{codebook_dir}{path to directory containing codebook yaml files}

\item{mapping_file}{path to yaml file with map from targets to REDCap variables}

\item{output_dir}{new directory where the resulting codebook yaml will be saved
(this will \emph{not} overwrite the existing codebook)}
}
\value{
side effect: writes yaml codebook to \code{output_path}. The path is also
invisibly returned.
}
\description{
Load redcap variable info using REDCap API and fill in codebook info for
related targets' vars.
}
