% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_find_file}
\alias{camr_find_file}
\title{Find Path to Most Recent File}
\usage{
camr_find_file(
  nm_desc,
  chr_path,
  chr_project = NULL,
  chr_date = NULL,
  chr_commit = NULL,
  chr_date_tolerance = "s",
  chr_type = "file",
  lgl_recurse = TRUE,
  tgt_depends = NULL
)
}
\arguments{
\item{nm_desc}{A character string, the file name
or part of the file name that can uniquely identify
it from other files.}

\item{chr_path}{A character string, the name of the
file path to search in.}

\item{chr_project}{An optional character string, the
name of the project.}

\item{chr_date}{An optional character string, the
date of the file.}

\item{chr_commit}{An optional character string, the
commit number of the file.}

\item{chr_date_tolerance}{An optional character string,
margin of error for the date of the file.
('s', 'm', 'h', 'd').}

\item{chr_type}{An optional character string,
file type to find.
('file', 'directory').}

\item{lgl_recurse}{Logical; if TRUE, the function
will recurse fully.}

\item{tgt_depends}{An optional target. Used to clarify dependencies in a
targets pipeline. In other words, if you are using
camr_find_file to produce a target in a targets pipeline,
and it depends on another target (e.g. the downloading of
REDCap data), you can indicate that target here so that
the targets algorithm can detect the dependency.}
}
\value{
A character string, an absolute path to a file or directory.
}
\description{
Function to find the absolute path to the most
up-to-date version of a file or directory given a
partial match to its name and the file location.
}
\author{
Michael Pascale, Bryn Evohr
}
