% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camr-shorthand.R
\name{camr_for}
\alias{camr_for}
\title{Shorthand to create a for loop wrapped in a progress bar.}
\usage{
camr_for(l, f)
}
\arguments{
\item{l}{A vector or list to loop over.}

\item{f}{A function to apply to each item.}
}
\description{
Shorthand to create a for loop wrapped in a progress bar.
}
\author{
Michael Pascale
}
