% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R22-yaml_codebook_maker.R
\name{camr_gen_codebook}
\alias{camr_gen_codebook}
\title{Create a codebook for a target dataframe using codebook yaml}
\usage{
camr_gen_codebook(df, yaml_dir)
}
\arguments{
\item{df}{Target dataframe. Output will contain a row for every variable
in this dataframe.}

\item{yaml_dir}{Directory containing yaml codebook files for your project}
}
\value{
A dataframe containing the codebook.
}
\description{
Matches variables in target dataframe with variable info stored in a directory
of yaml files. Outputs a dataframe with the names of all the variables and
codebook info (what instrument it's from, the meaning of its values, etc).
Also checks codebook-dataset consistency using \code{camr_check_codebook}.
}
\author{
Zach Himmelsbach
}
