% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R15-data_quality_rules_gen.R
\name{camr_gen_data_quality_rules}
\alias{camr_gen_data_quality_rules}
\title{Generate data quality rules based on REDCap metadata}
\usage{
camr_gen_data_quality_rules(
  path_to_api_token,
  output_path = "data_quality_rules.csv",
  excluded_vars = c("record_joindate", "record_joindate_utc", "record_timezone",
    "record_code")
)
}
\arguments{
\item{path_to_api_token}{path to a text file containing project API token}

\item{output_path}{path to a file where csv should be saved; will \emph{not}
overwrite an existing file. Defaults to "data_quality_rules.csv"}

\item{excluded_vars}{Character vector of REDCap vars that we do not want to
check. By default, we exclude the MyCap variables.}
}
\value{
side-effect: saves csv at \code{output_path}. Invisibly returns the
data quality rules csv for inspection
}
\description{
Write data quality rules based on REDCap Metadata. Values should not be
missing when they are 1) Required AND 2) reached by branching logic.
Produces a csv that can be uploaded to the data quality module in REDCap.f
}
\author{
Zach Himmelsbach
}
