% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R99-Legacy.R
\name{camr_get}
\alias{camr_get}
\title{Get an item fromm the global \code{camr} object.}
\usage{
camr_get(name)
}
\arguments{
\item{name}{The property to get.}
}
\description{
Get an item fromm the global \code{camr} object.
}
\author{
Michael Pascale
}
