% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R23-REDCap_var_processing.R
\name{camr_guarantee_fields}
\alias{camr_guarantee_fields}
\title{A function for guaranteeing all REDCap fields appear even if
no participant has required them yet}
\usage{
camr_guarantee_fields(
  df,
  target_form,
  api_token_file = Sys.getenv("API_TOKEN_FILE"),
  meta_data_df = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{target_form}{A character vector. The REDCap form names you want
to guarantee variables from. E.g. \code{target_form = "demographics"} ensures
all fields on the Demographics form are on the dataframe.}

\item{api_token_file}{A string. Path to a .txt file with the REDCap project
API token}

\item{meta_data_df}{A dataframe (optional). A dataframe containing variable
metadata for the REDCap project (produced by \link{camr_redcap_field_meta}). If
provided, \code{api_token_file} is not used to pull metadata and this is used
instead. This can help to avoid repeatedly pulling the project metadata.}
}
\value{
A dataframe with all REDCap project variables represented. Variables
that were missing will be all NA.
}
\description{
This function ensures that all REDCap project variables are represented as
columns in \code{df}. Missing columns are added to the returned dataframe.
}
