% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_in_codebook_entry}
\alias{camr_in_codebook_entry}
\alias{camr_ice}
\title{Specify Content for Codebook Entries}
\usage{
camr_in_codebook_entry(type, ...)

camr_ice(type, ...)
}
\arguments{
\item{type}{The type of formatting to carry out, where...
\itemize{
\item \code{content} Converts two vectors into a list
with \code{content} and \code{additional_content}.
\item \code{custom_abbr} Converts two vectors of (a) abbreviations
and (b) their associated labels into a matrix that
can passed to the \code{custom_category},
\code{custom_data_type}, or \code{custom_subcategory}
arguments.
\item \code{inventory} Takes output from the
\link{camr_inventories} function and converts it
into a list with \code{content} and \code{additional_content}.
\item \code{missing} Takes a vector of codes for missing
data and converts it into a character vector.
}}

\item{...}{Additional arguments to be formatted.}
}
\value{
Either (1) a named list with two elements, \code{content}
and \code{additional_content}, (2) a matrix with abbreviations
and associated labels, or (3) a character vector.
}
\description{
Convenience function to ensure inputs for
\link{camr_new_codebook_entry} and \link{camr_add_codebook_entry}
are correctly formatted.
}
\author{
Kevin Potter
}
