% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_instrument_event_map}
\alias{camr_instrument_event_map}
\title{Download REDCap mapping of instruments to events via API}
\usage{
camr_instrument_event_map(api_token_path)
}
\arguments{
\item{api_token_path}{path to txt file containing project API token}
}
\value{
A dataframe that maps instruments to events
}
\description{
Get the mapping of instruments to events as a dataframe. This function
takes a filepath to a txt file with a project API token and returns a
dataframe with mapping.
}
\author{
Zach Himmelsbach
}
