% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R06-Inventories_scales_questionnaires.R
\name{camr_inventories}
\alias{camr_inventories}
\title{Pre-defined Output for Known Scales and Subscales}
\usage{
camr_inventories(abbreviation = NULL, subscale = "")
}
\arguments{
\item{abbreviation}{The abbreviation for a scale, used to
look up the associated pre-defined output.}

\item{subscale}{The name of a subscale of the scale
given above, used to look up the relevant ancillary
output.}
}
\value{
A list consisting of...
\itemize{
\item \code{Description} A brief description of the
scale/subscale and its purpose
\item \code{Scale} A list with the scale name, abbreviation,
number of items, lowest to highest possible score (if
applicable), clinical cut-offs (if applicable), and
APA-styled references for the scale development/interpretation;
\item \code{Subscale} A list with the subscale name,
number of items, lowest to highest possible score,
clinical cut-offs (if applicable).
\item \code{Units} The units of measurement.
}
}
\description{
A convenience function that, given an abbreviation
for a scale, inventory, or questionnaire, returns
a list of information useful for data dictionary
purposes.
}
\examples{
# List all possible inputs (i.e., known
# scales, inventories, and questionnaires)
camr_inventories()

# Example input
camr_inventories( 'HADS' )
camr_inventories( 'HADS', subscale = 'Anxiety' )

}
\author{
Megan Cooke, Jacob Weiss, Kevin Potter
}
