% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_join_on_idx}
\alias{camr_join_on_idx}
\title{Safely left-join two dataframes on CAM index columns.}
\usage{
camr_join_on_idx(df_x, df_y)
}
\arguments{
\item{df_x}{A dataframe, with columns following the CAM naming convention.}

\item{df_y}{A dataframe, with columns following the CAM naming convention.}
}
\value{
left_join(df_x, df_y, by=camr_list_idx(df_y))
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
By convention, columns beginning with the prefix IDX are index columns. This
function will perform a \code{left_join}, mapping the IDX columns of \code{df_y} onto
those matching in \code{df_x}.
}
\examples{
\dontrun{
df_scans |>
  camr_join_on_idx(
    df_demographics)
  )
}
}
\keyword{internal}
