% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S00-Utilities.R
\name{camr_key}
\alias{camr_key}
\title{Securely Fetch a Key with \link{keyring}}
\usage{
camr_key(service, username = NULL, keyring = NULL)
}
\arguments{
\item{service}{Name of the service with which to associate the key.}

\item{username}{Optional. The username of the account to lookup. If NULL
(default), the name of the current project will be used as defined in the
\code{config.yml}. If empty string, \code{\link[keyring:key_get]{keyring::key_get()}} returns the first key
in the keyring matching the service.}

\item{keyring}{Optional. The name of the keyring on which the key is stored.
If NULL (default), the keyring specified in \code{config.yml} will be used.}
}
\description{
Wrapper around \code{\link[keyring:key_get]{keyring::key_get()}} with ability to pull keyring name from
the project \code{config.yml} and sensible defaults given the project structure.
}
\author{
Michael Pascale
}
