% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S06-FilesAndPaths.R
\name{camr_latest}
\alias{camr_latest}
\title{Fetch the latest version a file given a description.}
\usage{
camr_latest(
  prefix,
  description,
  extension,
  root = NULL,
  path = NULL,
  project = TRUE,
  recurse = TRUE
)
}
\arguments{
\item{prefix}{Passed to \code{camr_path}.}

\item{root}{Optional. Passed to \code{camr_path}.}

\item{path}{Optional. Base path of the file to be found.}

\item{project}{Optional. Whether to expect the project name to prefix the
file description field. Defaults to TRUE.}

\item{recurse}{Optional. Whether to search recursively. Defaults to TRUE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Expects files named in the format output by \code{camr_filename}.
}
\author{
Michael Pascale
}
\keyword{internal}
