% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_list_elements}
\alias{camr_list_elements}
\title{List Elements as Embedded in R Code}
\usage{
camr_list_elements(
  x,
  left = "  \\"",
  right = "\\",\\n",
  first = "c(\\n",
  last = ")",
  drop_last = ",",
  console = TRUE
)
}
\arguments{
\item{x}{A vector of values.}

\item{left}{A character string to attach
to the left of each element in \code{x}.}

\item{right}{A character string to attach
to the right of each element in \code{x}.}

\item{first}{A character string to attach
to the left of the final collapsed string.}

\item{last}{A character string to attach
to the right of the final collapsed string.}

\item{drop_last}{An optional character string
indicating any patterns to remove from the
final element.}

\item{console}{A logical value; if \code{TRUE}
displays the results in the console window,
otherwise returns a character string.}
}
\value{
Either a message to the console window
or a character string.
}
\description{
Function to generate R code listing the distinct
elements of \code{x} (e.g., embedding them in a
concatenate or list statement).
}
\examples{
camr_list_elements( LETTERS[1:3] )

# Operates only on distinct values
camr_list_elements( rep( LETTERS[1:3], each = 2 ) )

# Can adjust how elements are embedded in R code
camr_list_elements( 1:3, left = '  ', right = ',\n' )
camr_list_elements( LETTERS[1:3], first = 'list(\n' )

}
