% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_list_idx}
\alias{camr_list_idx}
\title{List CAM index columns of a dataframe.}
\usage{
camr_list_idx(df_x)
}
\arguments{
\item{df_x}{A dataframe, with columns following the CAM naming convention.}
}
\value{
A character vector of the names in df_x prfixed by IDX.
}
\description{
By convention, columns beginning with the prefix IDX are index columns.
}
