% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_make_index}
\alias{camr_make_index}
\title{Create Numerical Index for Variable}
\usage{
camr_make_index(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments for the
\link{camr_missing} function.}
}
\value{
A vector of integers starting from 1
indexing the distinct entries in the vector
\code{x}.
}
\description{
Creates a numerical index starting from 1
going to the total number of distinct entries
for a variable.
}
\examples{
camr_make_index( rep( LETTERS[1:3], each = 2 ) )

}
\author{
Kevin Potter
}
