% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R14-Yaml_codebook.R
\name{camr_make_yaml_codebook}
\alias{camr_make_yaml_codebook}
\title{Generate and fill codebook for current project (wrapper around suite of
codebook functions)}
\usage{
camr_make_yaml_codebook(api_token_path, output_dir)
}
\arguments{
\item{api_token_path}{Character. Path to a txt file containing the project
API token}

\item{output_dir}{Character. Directory where the yaml files will be stored}
}
\value{
Side-effect: Populates \code{output_dir} with one yaml file per targets
dataframe. Invisibly returns \code{output_dir}
}
\description{
Create codebook based on existing targets in the current project. This function
creates a directory with yaml files for each target dataframe. Within each
file, there's information about the dataframe as well as entries for each
variable within that dataframe. The function maps variables to their REDCap
source variables. For variables with 1-to-1 mappings, much of the variable
info is filled in automatically. The function also outputs the blank codebook
and the mapping file from REDCap vars to target dataframe vars. If the mappings
need to be updated manually, \code{camr_fill_codebook_from_redcap} can be used to
refill the codebook.
}
\author{
Zach Himmelsbach
}
