% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R14-Yaml_codebook.R
\name{camr_map_redcap_vars}
\alias{camr_map_redcap_vars}
\title{Scrape functions that generate targets to map variable names to the
REDCap variables they were created from}
\usage{
camr_map_redcap_vars(
  output_dir,
  api_token_path,
  exclude_redcap_vars = "co",
  src_dir = "src",
  exclude_targets = c("rds_download", "chr_path_redcap_data", "chr_path_tlfb_data",
    "lst_redcap_data", "df_tlfb_raw", "df_id_table", "df_redcap_raw")
)
}
\arguments{
\item{output_dir}{file path to directory where the yaml should be saved}

\item{api_token_path}{file path to txt file containing redcap project token}

\item{exclude_redcap_vars}{character vector of redcap variables the parser
should \emph{not} search for. Default is "co" because it captures tons of stuff.
(I have since refactored the regex so this isn't required, but leaving the
option for now.)}

\item{src_dir}{directory to search for functions that generate variables.
Default is "src/".}

\item{exclude_targets}{a character vector with names of targets to exclude.
Excludes core targets by default.}
}
\value{
side effect: a yaml file is created in \code{output_dir}. Invisibly returns
the list object containing the mapping
}
\description{
Scrape the functions that generate targets and map final variable names
to their REDCap ancestor variables. Produces a yaml file that lists the
target dataframes with their mappings (from final vars to REDCap vars). The
scraper looks for transmute calls. If a variable is constructed from more than
one REDCap ancester, the mapping is listed as "composite".
}
\author{
Zach Himmelsbach
}
