% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R09-Miscellaneous.R
\name{camr_meddra_join}
\alias{camr_meddra_join}
\title{Join MedDRA Terms}
\usage{
camr_meddra_join(df, code_col, lst_meddra)
}
\arguments{
\item{df}{The dataframe.}

\item{code_col}{The column name with MedDRA codes.}

\item{lst_meddra}{The MedDRA database list.}
}
\value{
Dataframe joined with MedDRA terms and levels.
}
\description{
Function to add MedDRA terms and levels
based on the corresponding MedDRA code
in a dataframe.
}
\examples{
data(lst_meddra)

camr_meddra_join(df_medhx, INV.INT.MedHx.MedDRA.Code,
  lst_meddra) |>
  mutate(
    INV.CHR.MedHx.MedDRA.Term = code_name,
    INV.CHR.MedHx.MedDRA.HLT.Term = hlt_name,
    INV.CHR.MedHx.MedDRA.HLGT.Term = hlgt_name,
    INV.CHR.MedHx.MedDRA.SOC.Term = soc_name,
    INV.FCT.MedHx.MedDRA.Level = level,
    .after = INV.INT.MedHx.MedDRA.Code
  ) |>
  select(-c(level, code_name, hlt_name,
    hlgt_name, soc_name)) -> df_medhx_meddra

}
\author{
Bryn Evohr
}
