% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_merge_data_frames}
\alias{camr_merge_data_frames}
\title{Merge Multiple Data Frames Together}
\usage{
camr_merge_data_frames(...)
}
\arguments{
\item{...}{The data frames to merge together.}
}
\value{
A single data frame.
}
\description{
Function to merge multiple data frames
(assuming equal number of rows) together.
}
\examples{
dtf_1 <- data.frame( X = 1:2 )
dtf_2 <- data.frame( Y = LETTERS[1:2] )
dtf_3 <- data.frame( Z = 3:4 )

print( camr_merge_data_frames( dtf_1, dtf_2, dtf_3 ) )

}
