% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_missing}
\alias{camr_missing}
\title{Identify Missing Data}
\usage{
camr_missing(
  x,
  codes = NULL,
  attribute_label = "missing",
  return_codes = FALSE
)
}
\arguments{
\item{x}{A vector of values.}

\item{codes}{A vector of possible values
of \code{x} that indicate missing data.}

\item{attribute_label}{The label for
sublist containing codes for missing data
in the list of attributes for \code{x}.}

\item{return_codes}{A logical value; if
\code{TRUE} returns the set of codes
for missing data actually found in the
data set.}
}
\value{
A logical vector equal to \code{TRUE}
if a value is missing and \code{FALSE} otherwise.
}
\description{
Function that checks for missing values in
a vector. Allows for checking against multiple
different codes that denote missing values
(e.g., a character vector where missing
values can be either \code{NA} or \code{""}).
}
\examples{
x <- c( NA, "", "Cat", "Dog", "Missing" )
print( camr_missing(x) )

# Custom codes
print( camr_missing(x, codes = c( NA, "Missing" ) ) )

# Codes can be saved as an attribute
x <- x |> camr_add_attribute( c( NA, "Missing" ), "missing" )
# Function can automatically access
print( camr_missing(x) )

# Can return only applicable codes for a vector
x <- c( "", "Cat", "Dog" ) # No NA values
print( camr_missing(x, return_codes = TRUE ) ) # Only "" returned

}
\author{
Kevin Potter
}
