% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_new_codebook_entry}
\alias{camr_new_codebook_entry}
\title{Create New Codebook Entry}
\usage{
camr_new_codebook_entry(
  Variable_name = "",
  Category = "",
  Data_type = "",
  Subcategory = "",
  Description = "",
  Values_and_labels = "",
  Inventory = "",
  Units_of_measurement = "",
  Descriptive_summary = "",
  Codes_for_missing_data = "",
  Deidentified = "FALSE",
  Collected_over = "",
  Created_from = "",
  Data_quality_checks = "",
  Notes = ""
)
}
\arguments{
\item{Variable_name}{A character string, the column
name for the variable of interest.}

\item{Category}{The label associated with the 3-letter
abbreviation for the variable's category (see
\link{camr_column_abbreviations}).}

\item{Data_type}{The label associated with the 3-letter
abbreviation for the variable's data type (see
\link{camr_column_abbreviations}).}

\item{Subcategory}{The label associated with the custom
abbreviation for the variable's subcategory (see
\link{camr_column_abbreviations}).}

\item{Description}{A character string, a brief human-readable
description of what the variable is.}

\item{Values_and_labels}{A named list of two vectors,
(1) \code{content} containing the distinct values for
the variable, and (2) \code{additional_content} containing
the label associated with each value in (1).}

\item{Inventory}{A named list of two vectors,
(1) \code{content} containing the details for
the inventory, scale, or questionnaire, and (2)
\code{additional_content} containing the label
associated with each detail given in (1).}

\item{Units_of_measurement}{A character string, the
units of measurement for the variable.}

\item{Descriptive_summary}{A named list of two vectors,
(1) \code{content} containing summary statistics for
the variable, and (2) \code{additional_content} containing
the label associated with each statistic in (1).}

\item{Codes_for_missing_data}{A character vector, the
codes (surrounding by back-ticks) for missing data.}

\item{Deidentified}{Either \code{FALSE} if the variable
is not deidentified (i.e., contains potential patient
health information) or \code{TRUE} otherwise.}

\item{Collected_over}{A named list of two vectors,
(1) \code{content} containing distinct values for
groups, time points, etc., over which the variable was
collected, and (2) \code{additional_content} containing
the column name associated with the values in (1).}

\item{Created_from}{A character vector, the original
REDCap variables from which the variable was created.}

\item{Data_quality_checks}{A character string, notes
on any data quality checks relevant to the variable.}

\item{Notes}{A character vector, any additional notes
to include in the codebook entry.}
}
\value{
A data frame of the additional class
\code{codebook_entry}.
}
\description{
Function to create a data frame of additional
class \code{codebook_entry}.
}
\details{
The use of commas should be avoided for
codebook entries in order to ensure easy conversion
to a .csv file.
}
\author{
Kevin Potter
}
