% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_pass}
\alias{camr_pass}
\title{Do Nothing}
\usage{
camr_pass(
  ...,
  .lgl_pass_quietly = TRUE,
  .chr_pass_message = "(camr::camr_pass) not implemented."
)
}
\arguments{
\item{.chr_pass_message}{Optional. Message to display if warnings enabled.}

\item{.lgl_pass_quetly}{Optional. If true a warning message is displayed.}
}
\description{
This function provides functionality similar to Python's \code{pass} keyword. It
will ignore all input and return NULL.
}
\examples{
\dontrun{
tryCatch({
 # non-critical, ignorable error.
}, error=camr_pass)

fn_example <- function(x, y) {
 camr_pass(.lgl_pass_quietly=FALSE)
}
}
}
\author{
Michael Pascale
}
