% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R10-Experimental.R
\name{camr_path}
\alias{camr_path}
\title{Construct a Path Given Base Paths in Project Configuration File}
\usage{
camr_path(
  prefix = c("input", "output"),
  root = NULL,
  path = NULL,
  real = TRUE,
  create = FALSE
)
}
\arguments{
\item{prefix}{'input' or 'output' corresponding to the prefix of the
corresponding parameter in \code{config.yml}.}

\item{root}{Optional. Name of the base path to lookup.}

\item{path}{Optional. Path tail to concatenate at the end of the base path.}

\item{real}{Optional. Whether to create a directory at the generated path.
Cannot be used with \code{real}. Defaults to \code{FALSE}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Expects the project \code{config.yml} to contain a named list of base paths,
\verb{(?<prefix>\\w+)-root}, and a default base path name,
\verb{(?<prefix>\\w+)-root-default}.
}
\author{
Michael Pascale
}
\keyword{internal}
