% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{camr_paths}
\alias{camr_paths}
\title{Construct and check the existence of paths given the prefix directory.}
\usage{
camr_paths()
}
\description{
Construct and check the existence of paths given the prefix directory.
}
\author{
Michael Pascale
}
