% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_pivot_redcap_eav}
\alias{camr_pivot_redcap_eav}
\title{Extract tables from REDCap EAV downloads}
\usage{
camr_pivot_redcap_eav(
  df_redcap_eav,
  vchr_fields,
  lgl_collapse_checkboxes = FALSE,
  chr_event_pattern = NULL,
  chr_field_column = "field_name",
  chr_value_column = "value",
  chr_event_column = "redcap_event_name"
)
}
\arguments{
\item{df_redcap_eav}{A dataframe with REDCap data in EAV format.}

\item{vchr_fields}{A set of REDCap field names to extract.}

\item{lgl_collapse_checkboxes}{Whether to collapse checkbox fields with '|'.}

\item{chr_event_pattern}{Optional. A pattern to validate event names against.}

\item{chr_field_column}{Optional. Defaults to the API provided "field_name".}

\item{chr_value_column}{Optional. Defaults to the API provided "value".}

\item{chr_event_column}{Optional. Defaults to the API provided "redcap_event_name".}
}
\value{
A dataframe, with the columns of df_redcap_eav and those specified by
vchr_fields.
}
\description{
\code{camr_download_redcap()} downloads data in entity-attribute-value (EAV)
format (one row per datapoint).
}
\keyword{internal}
