% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camr-globals.R
\name{camr_popd}
\alias{camr_popd}
\title{Pop the last path from the directory stack and switch into it.}
\usage{
camr_popd()
}
\description{
Pop the last path from the directory stack and switch into it.
}
\author{
Michael Pascale
}
