% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S00-Globals.R
\name{camr_prefix}
\alias{camr_prefix}
\title{Set a directory prefix for data files.}
\usage{
camr_prefix(prefix = getwd())
}
\arguments{
\item{prefix}{Optional. A path to the desired data directory. Defaults to the
working directory.}
}
\description{
Set a directory prefix for data files.
}
\author{
Michael Pascale
}
