% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Pretend.R
\name{camr_pretend}
\alias{camr_pretend}
\title{This is a pretend function}
\usage{
camr_pretend(v)
}
\arguments{
\item{v}{A character vector}
}
\value{
Side effect: prints a nice message
}
\description{
Totally fake; for educational purposes only
}
