% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R23-REDCap_var_processing.R
\name{camr_process_redcap_var}
\alias{camr_process_redcap_var}
\title{Function to process individual variables in a script}
\usage{
camr_process_redcap_var(x, api_token_path = Sys.getenv("API_TOKEN_PATH"))
}
\arguments{
\item{redcap_varname}{A string. The name of the redcap variable.}

\item{api_token_path.}{A string. Path to a .txt file containing the redcap
project API token.}
}
\value{
A vector of the variable processed according to \code{camr_process_redcap_vars}.
}
\description{
This function applies the same processing as \code{camr_process_redcap_vars},
but allows it to be called on a single variable, and just returns the data
(so it can be used e.g. inside calls to \code{dplyr::mutate}).
It also skips the use of automated naming.
Note: this does NOT work for checklist variables.
}
