% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R23-REDCap_var_processing.R
\name{camr_process_redcap_vars}
\alias{camr_process_redcap_vars}
\title{Process REDCap variables into standard CAM variables}
\usage{
camr_process_redcap_vars(
  df,
  prefix = "INV",
  api_token_path = Sys.getenv("API_TOKEN_FILE"),
  custom_form_name = NULL
)
}
\arguments{
\item{df}{A dataframe. The column names must follow the CAM REDCap varname
convention (\code{form_name}_\code{fieldname}).}

\item{prefix}{A string. The first part of the new variable names. Usually
this will be "INV" (the default), but sometimes you will want to use "SBJ".}

\item{custom_form_name}{A string. A custom form name to be used in place
of the automatically derived form name
(useful when e.g. combining multiple forms, like constructing
visit info from the start of visit AND end of visit forms). If \code{NULL}, the form
name is automatically derived from the REDCap variable name.}

\item{api_token_path.}{A string. The path to your REDCap project API token.
You can add it to your project .Renviron file as "API_TOKEN_FILE" and the
default will pick it up automatically.}
}
\value{
A dataframe with variables processed and renamed according to CAM
standards.
}
\description{
Takes in a dataframe where column names are REDCap variable names. Processes
data according to field metadata (e.g. Dropdown and Radio variables are
converted to factors). Only variables that follow the REDCap variable convention
(i.e. they contain "_") will be processed. Other variables (e.g. IDS.CHR.Subject),
will be left alone. Checklist variables are split into multiple indicator variables.
}
