% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S06-FilesAndPaths.R
\name{camr_processed_data_to_csv}
\alias{camr_processed_data_to_csv}
\title{Convert Processed Data to .csv Files}
\usage{
camr_processed_data_to_csv(
  dtf,
  proj,
  desc,
  cur_time = "",
  commit_num = NULL,
  file_path = ""
)
}
\arguments{
\item{dtf}{A data frame with column attributes that
include a list of class \code{dictionary_meta_data}.}

\item{proj}{A character string, the abbreviation for
the study (e.g., 'MMJ', 'ARCHES', 'PCORI', etc.).}

\item{desc}{A character string, the description of
the file (e.g., 'Processed_data').}

\item{cur_time}{The date and time the file was created,
preferably in the format YYYY_MM_DD-HH_MM.
If not provided, computed automatically via the base
R function \code{\link[base:Sys.time]{base::Sys.time()}}.}

\item{commit_num}{A character string, the 7-digit alphanumeric
commit reference - if not provided, automatically computed.}

\item{file_path}{An optional character string, the absolute
path to the folder to save the files to - otherwise, saves
to the current working directory.}
}
\value{
Creates two .csv files, one for the processed data
and another for the associated data dictionary (created
automatically via the \code{\link[=data_frame_from_dictionary_meta_data]{data_frame_from_dictionary_meta_data()}}
function).
}
\description{
Convenience function that takes a data frame with
processed data and converts it to a .csv file along
with an additional .csv file with a data dictionary.
}
