% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_pull_codebook_entry}
\alias{camr_pull_codebook_entry}
\alias{camr_puce}
\title{Extract Codebook Entry for a Variable}
\usage{
camr_pull_codebook_entry(x = NULL, variable = "")

camr_puce(...)
}
\arguments{
\item{x}{Either (a) a data frame or (b) a column
in a data frame.}

\item{variable}{If \code{x} is a data frame, the column
in the data frame from which the codebook entry
should be extracted.}
}
\value{
Either (a) the data frame of additional
class \code{codebook_entry} for the specified
variable, or (b) a logical vector equal to
\code{TRUE} for columns with codebook entries.
}
\description{
Function to extract a codebook entry for
a specified variable, or indicate all
variables in a data frame that have
codebook entries.
}
\examples{
data( "example_CAM_data_set" )
dtf <- example_CAM_data_set[,1:2]

# Add codebook entry to data frame
dtf <- dtf |>
  camr_add_codebook_entry(
    IDS.INT.Screening,
    description = "Identifier assigned at screening session"
)
# Identify columns with and without codebook entries
camr_pull_codebook_entry( dtf )

# Extract codebook entry
dtf |> camr_pull_codebook_entry( IDS.INT.Screening )
# Alternative approach
camr_pull_codebook_entry( dtf$IDS.INT.Screening )

}
\author{
Kevin Potter
}
