% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S00-Globals.R
\name{camr_pushd}
\alias{camr_pushd}
\title{Push a path to the directory stack.}
\usage{
camr_pushd(path = getwd())
}
\arguments{
\item{path}{Optional. The path to save. Defaults to the working directory.}
}
\description{
Push a path to the directory stack.
}
\author{
Michael Pascale
}
