% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R09-Miscellaneous.R
\name{camr_read_tlfb_v2}
\alias{camr_read_tlfb_v2}
\title{Read in TLFB V1/V2 JSON Data}
\usage{
camr_read_tlfb_v2(chr_path_to_json_dir, recursive = FALSE)
}
\arguments{
\item{chr_path_to_json_dir}{Directory path containing JSON files.}

\item{recursive}{Whether to search subfolders of the directory for data files.}
}
\value{
A dataframe.
}
\description{
Read in TLFB V1/V2 JSON Data
}
