% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_redcap_download}
\alias{camr_redcap_download}
\title{Title}
\usage{
camr_redcap_download(chr_rc_uri = "", chr_rc_token = "")
}
\arguments{
\item{chr_rc_uri}{...}

\item{chr_rc_token}{...}
}
\value{
A list.
}
\description{
Description.
}
\examples{
# Examples

}
