% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_redcap_event_data}
\alias{camr_redcap_event_data}
\title{Pull Event info from REDCap project process it into a dataframe}
\usage{
camr_redcap_event_data(api_token = Sys.getenv("API_REDCAP_TOKEN"))
}
\arguments{
\item{api_token}{A string. The API token for the REDCap project}
}
\value{
A dataframe
}
\description{
Pull Event data using REDCap API and convert it into a
convenient data frame. Used to automate processing of visit
data
}
