% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_redcap_project_info}
\alias{camr_redcap_project_info}
\title{Download REDCap project info via API}
\usage{
camr_redcap_project_info(api_token_path)
}
\arguments{
\item{api_token_path}{path to txt file containing project API token}
}
\value{
A list of REDCap project info
}
\description{
Get the project info for a REDCap project. This information includes project
ID, title, creation/production time, in_production flag, longitudinal flag,
and other info. This function takes a filepath to a txt file with a project
API token and returns a list.
}
\author{
Zach Himmelsbach
}
