% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camr-globals.R
\name{camr_reset}
\alias{camr_reset}
\title{Reset the global cache object. Delete the \code{.cache} file.}
\usage{
camr_reset()
}
\description{
Reset the global cache object. Delete the \code{.cache} file.
}
\author{
Michael Pascale
}
