% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{camr_save}
\alias{camr_save}
\title{Save the global cache object to disk.}
\usage{
camr_save()
}
\description{
The global cache is stored to a \code{.cache} Rds in the working directory.
}
\author{
Michael Pascale
}
