% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R99-Legacy.R
\name{camr_set}
\alias{camr_set}
\title{Set a miscellaneous item within the global \code{camr} object and save to disk.}
\usage{
camr_set(name, value)
}
\arguments{
\item{name}{The property to set.}

\item{value}{The value to set.}
}
\description{
Set a miscellaneous item within the global \code{camr} object and save to disk.
}
\author{
Michael Pascale
}
