% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Wrangling.R
\name{camr_shuffle_groups}
\alias{camr_shuffle_groups}
\title{Shuffle Grouping Variable Levels}
\usage{
camr_shuffle_groups(
  dtf,
  id,
  group,
  within = NULL,
  include = NULL,
  group_levels = NULL,
  original_freq = FALSE,
  save_unshuffled = TRUE,
  rng_seed = NULL
)
}
\arguments{
\item{dtf}{A data frame.}

\item{id}{A character string, the column with participant
identifiers.}

\item{group}{A character string, the column for the
grouping variable.}

\item{within}{An optional character string, the column with
the levels for a within-participant variable (e.g.,
time points or visits).}

\item{include}{An optional logical vector matching in length
to the number of rows in \code{dtf}, indicating the subset
of cases to shuffle. If \code{NULL} all rows are used.}

\item{group_levels}{An optional character vector, the subset
of levels of \code{group} to consider.}

\item{original_freq}{Logical; if \code{TRUE} shuffles
data in a way that preserves the original frequencies
for group levels - otherwise, assigns new levels in
equal frequencies.}

\item{save_unshuffled}{Logical; if \code{TRUE} adds a
new column with the original unshuffled group levels.}

\item{rng_seed}{An integer, the RNG seed to use to ensure
reproducibility.}
}
\value{
A data frame with shuffled group levels for the
  \code{group} variable.
}
\description{
Function that shuffles the levels of a grouping variable
(e.g., treatment or intervention assignments) over
participants (and optionally a within-participant
variable like study visit or time point). This is
useful, for example, to create a data set for an
analyst-blind design.
}
\author{
Kevin Potter
}
