% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_by_ID}
\alias{camr_SWA_linking_code_by_ID}
\title{Linkage Patterns by Linked Identifier}
\usage{
camr_SWA_linking_code_by_ID(dtf_linked, lgc_update = FALSE)
}
\arguments{
\item{dtf_linked}{A data frame, output from
the \link{camr_SWA_linking_code}
function. Must have the columns
\code{IDX.CHR.Linked.ID} and
\code{SSS.INT.Time_point}.}
}
\value{
A wide-form data frame with one row
per linked ID along with the pattern of linked
time points (e.g., \code{'0-1'} means a link
from the baseline to first time point).
}
\description{
Function to determine the pattern of
linked time points for each linked ID.
}
