% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_items}
\alias{camr_SWA_linking_code_items}
\title{Extract Linking Item Column Names}
\usage{
camr_SWA_linking_code_items(dtf_linked)
}
\arguments{
\item{dtf_linked}{A data frame, output from
the \link{camr_SWA_linking_code}
function.}
}
\value{
A character vector, the variables used
as linking items.
}
\description{
Function to extract column names for all
variables used as linking items.
}
