% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_performance}
\alias{camr_SWA_linking_code_performance}
\title{Summary of Linking of Records}
\usage{
camr_SWA_linking_code_performance(
  dtf_linked,
  lst_groups = list(Time = "SSS.INT.Time_point"),
  lgc_display = TRUE
)
}
\arguments{
\item{dtf_linked}{A data frame, the output
from \link{camr_SWA_linking_code}.}

\item{lst_groups}{A named list of column names,
the grouping factors to consider when summarizing
the number of records linked.}

\item{lgc_display}{A logical value; if \code{TRUE}
prints the results to the console in addition
to returning output as a list.}
}
\value{
A list of data frames.
}
\description{
Function to summarize the performance of
the linking code for the school-wide
assessment. Provides summary statistics
for what records were linked, and if
columns with the 'true' IDs are detected,
provides details on hits and correct
rejections.
}
\examples{
# Linking across time points
dtf_long <- camr_SWA_linking_code_simulate('demo')
dtf_linked <- camr_SWA_linking_code(dtf_long)
lst_summary <- camr_SWA_linking_code_performance(
  dtf_linked, lgc_display = FALSE
)

}
