% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_rows}
\alias{camr_SWA_linking_code_rows}
\title{Extract Rows Identified as Possible Links}
\usage{
camr_SWA_linking_code_rows(dtf_linked, int_row = 1, lgc_filter = TRUE)
}
\arguments{
\item{dtf_linked}{A data frame, output from
the \link{camr_SWA_linking_code}
function.}

\item{int_row}{The row from \code{dtf_linked} to
process.}

\item{lgc_filter}{A logical value; if \code{TRUE}
returns the data frame rows, otherwise
returns the extract rows.}
}
\value{
Either a data frame or a vector of row indices.
}
\description{
Extract rows identified as possible links for a
given record.
}
