% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_select}
\alias{camr_SWA_linking_code_select}
\title{Select Columns for Linking Items and Useful Information}
\usage{
camr_SWA_linking_code_select(dtf_linked, chr_extra = "")
}
\arguments{
\item{dtf_linked}{A data frame, output from
the \link{camr_SWA_linking_code}
function.}

\item{chr_extra}{An optional character vector,
additional columns to include.}
}
\value{
A data frame,
}
\description{
Extracts columns with row indices, time points,
linked IDs, and potential linkable rows, along
with linking item variables (assuming standard
column names).
}
