% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R12-SWA_linking_code.R
\name{camr_SWA_linking_code_simulate}
\alias{camr_SWA_linking_code_simulate}
\title{Simulate Data to Link}
\usage{
camr_SWA_linking_code_simulate(chr_type = "debug")
}
\arguments{
\item{chr_type}{A character string, the type of
data set to generate, where \code{'debug'}
generates data for debugging purposes,
\code{'demo'} generates a small data set for
demonstration purposes,
\code{'accuracy'} generates data for accuracy
assessment, \code{'duplicate'} generates data
for a single time point with duplicate records,
and \code{'demo'} generates a small data set
for demonstration purposes.}
}
\value{
A data frame.
}
\description{
Function to simulate different types of data
to pass to the linking code function.
Provides (a) data for debugging purposes, (b)
realistic data for accuracy assessment, (c)
data at a single time point with duplicate
records, and (d) a small data set for
demonstration purposes.
}
\examples{
dtf_demo <- camr_SWA_linking_code_simulate( 'demo' )

}
\author{
Kevin Potter
}
