% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S07-FunctionDocumentation.R
\name{camr_template_function_documentation}
\alias{camr_template_function_documentation}
\title{Create a Template for a Function's Documentation}
\usage{
camr_template_function_documentation(func_name, script_path, tag_for_proj_fun)
}
\arguments{
\item{func_name}{A character string, the
name of the function for which a template
of the documentation is needed.}

\item{script_path}{A character string, the
path to the R script containing the function
(relative to the working directory).}

\item{tag_for_proj_fun}{A character string,
the substring that precedes every project
function. For example, if project functions
follow the format \code{project.RXX.function_name}
the input should be \code{project.R}.}
}
\value{
A message displayed in the console
with the template for the function
documentation, which can be copied and
pasted as needed.
}
\description{
Given an R script that contains a function whose
name and arguments follow a specific naming
convention, generates a template for the
function documentation (determining dependencies
on R packages, target outputs, and other project
functions).
}
