% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_unique_values}
\alias{camr_unique_values}
\title{Extract Unique Values}
\usage{
camr_unique_values(
  dtf,
  index,
  variables,
  default = NA,
  propagate = TRUE,
  first = TRUE,
  warn_if_multiple = TRUE
)
}
\arguments{
\item{dtf}{A data frame.}

\item{index}{The column in \code{dtf} that serves
as the grouping factor (i.e., each distinct case
in \code{index} should correspond to a unique value).
Non-standard evaluation possible.}

\item{variables}{A character vector, all columns in
\code{dtf} in which unique values should be extracted.}

\item{default}{The default value to return if unique
values are missing.}

\item{propagate}{A logical value; if \code{TRUE},
propagates unique values by \code{index} to match
the number of rows of \code{dtf}. If \code{FALSE}
returns the unique value for each distinct
case of \code{index}.}

\item{first}{A logical value; if \code{TRUE} uses the
first value found there is more than one value
identified for a given case in \code{index},
otherwise uses the final value.}

\item{warn_if_multiple}{A logical value; if \code{TRUE}
returns a warning whenever multiple values are
found for a given case in \code{index}.}
}
\value{
A vector of unique values associated with the
distinct cases in \code{index}.
}
\description{
Function that filters out missing values and
propagates the unique values associated with
distinct cases in an index column. Useful,
for example, taking values (like participant
characteristics) that were collected
at a single time point and propagating them
across all time points.
}
\examples{
data( example_CAM_data_set )

example_CAM_data_set |>
  camr_unique_values(
    IDS.INT.Screening,
    "SBJ.INT.Age"
 )

example_CAM_data_set |>
  camr_unique_values(
    IDS.INT.Screening,
    "SBJ.INT.Age",
    propagate = FALSE
  )

# Can extract values over multiple columns
example_CAM_data_set |>
  camr_unique_values(
    IDS.INT.Screening,
    c( "SBJ.INT.Age", "SBJ.INT.Age_at_visit_1" ),
    propagate = FALSE
  )
# Appropriately generates warning for
# 3rd individual who has different ages
# recorded across variables

}
\author{
Kevin Potter
}
