% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{camr_update_codebook_entry}
\alias{camr_update_codebook_entry}
\title{Update Codebook Entries for Data Frame}
\usage{
camr_update_codebook_entry(dtf)
}
\arguments{
\item{dtf}{A data frame whose variables have
codebook entries.}
}
\value{
A data frame with updated codebook entries.
}
\description{
Function that updates the descriptive summaries,
codes for misssing data, and cases collected over
for relevant codebook entries for a data frame
that is a subset of a larger data set.
}
