% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_util_collapse}
\alias{camr_util_collapse}
\title{Collapse Checkbox Values}
\usage{
camr_util_collapse(x)
}
\arguments{
\item{x}{An object containing the values to collapse.}
}
\description{
Collapse checkbox values when pivoting
REDCap raw data.
}
\examples{
pivot_wider(
  df_redcap_raw,
  names_from=field_name,
  values_from=value,
  values_fn=camr_util_collapse
)
}
\author{
Michael Pascale
}
\keyword{internal}
