% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Data_wrangling.R
\name{camr_values_labels}
\alias{camr_values_labels}
\title{Display Values and Associated Labels}
\usage{
camr_values_labels(dtf, values, labels)
}
\arguments{
\item{dtf}{A data frame.}

\item{values}{A character string, the column name for
the values of interest (non-standard evaluation possible).}

\item{labels}{A character string, the column name for
the labels of interest (non-standard evaluation possible).}
}
\value{
A data frame with a column for values and
a column for associated labels.
}
\description{
A function that takes two columns in a data frame
(assumed to be an initial column of values and
a subsequent column of associated labels)
and displays the assignment of values to labels.
}
\examples{
data( example_CAM_data_set )

example_CAM_data_set$SBJ.CHR.Biological_sex <-
  c( "Male", "Female" )[ example_CAM_data_set$SBJ.INT.Biological_sex ]

example_CAM_data_set |>
  camr_values_labels( SBJ.INT.Biological_sex, SBJ.CHR.Biological_sex )

}
\author{
Kevin Potter
}
