% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R20-Visit_and_dropout_flow.R
\name{camr_visit_and_dropout_flow}
\alias{camr_visit_and_dropout_flow}
\title{Visualize Visit flow, separately, for active participants and dropouts}
\usage{
camr_visit_and_dropout_flow(
  df_visits,
  df_status,
  output_dir = NULL,
  visit_flow_title = "Active/Completed Participants: Most Recent Completed Visit",
  dropout_flow_title = "Dropped Participants: Most Recent Completed Visit",
  dropout_statuses = c("LTFU", "Withdrawn", "Terminated"),
  label_max = 5
)
}
\arguments{
\item{df_visits}{A dataframe following the conventions for \code{df_visits} in the
standard pipeline. The function expects SSS.LGL.Visit.Completed to indicate
visit completion, IDX.INT.VisitNumber to provide the ordered visit numbers, and
SSS.FCT.VisitName to provide visit names (with levels that match the visit numbers).}

\item{df_status}{A dataframe following the conventions for \code{df_status} in the
standard pipeline. Expects dropouts to have \code{SBJ.FCT.Status} of
"LTFU", "Withdrawn", or "Terminated".}

\item{output_dir}{(Optional) A string of the directory to save the plots as
svg files.}

\item{visit_flow_title}{A string with the title of the visit flow plot.
Defaults to "Active/Completed Participants: Most Recent Completed Visit"}

\item{dropout_flow_title}{A string with the title of the dropout flow plot.
Defaults to "Dropped Participants: Most Recent Completed Visit"}
}
\value{
A list of ggplot2 objects. Optionally saves .svg to \code{output_path}
}
\description{
Create a bar plot that shows the most recent completed visit for participants
who are active (or completed). Takes the standard \code{df_visits} as input.
}
\author{
Zach Himmelsbach
}
