% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{camr_write_csv}
\alias{camr_write_csv}
\title{Write a CSV file with the standard options.}
\usage{
camr_write_csv(
  df_x,
  chr_filename,
  row.names = FALSE,
  na = "",
  lgl_literal = FALSE
)
}
\arguments{
\item{df_x}{A dataframe to be written.}

\item{chr_filename}{A dataframe, with columns
following the CAM naming convention.}

\item{...}{Additional arguments to write.csv(),}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
The provided filename will be expanded with \code{camr_name_file}.
}
\keyword{internal}
