% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R14-Yaml_codebook.R
\name{camr_yaml_codebook}
\alias{camr_yaml_codebook}
\title{Create a set of yaml codebooks for variables in targets defined
by _targets.R}
\usage{
camr_yaml_codebook(
  codebook_dir = "codebook",
  max_unique = 20,
  exclude_targets = c("rds_download", "chr_path_redcap_data", "chr_path_tlfb_data",
    "lst_redcap_data", "df_tlfb_raw", "df_id_table", "df_redcap_raw")
)
}
\arguments{
\item{codebook_dir}{Character. Directory to store the codebook files. This
must \emph{not} currently exist. Default "codebook".}

\item{max_unique}{Integer. If an integer or character variable has less than
\code{max_unique} distinct values, those values are listed in the codebook. Default 20.}

\item{exclude_targets}{a character vector with names of targets to exclude.
Excludes core targets by default.}
}
\value{
side effect: \code{codebook_dir} directory is populated with yaml codebooks
for each targets object. Invisibly returns list of variables for which
standard pipeline entries were \emph{not} found
}
\description{
Copy codebook for variables in the standard pipeline and generate stub entries
for new variables. Creates a directory in your project called "codebook" with
one .yaml file per target in _targets.R. Some information is automatically
inferred for new variables (data type, unique values, etc)
}
\author{
Zach Himmelsbach
}
