% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camrprojects-package.R
\docType{package}
\name{camrprojects-package}
\alias{camrprojects-package}
\alias{camrprojects}
\title{Portable Functions for R Projects at CAM-MGB}
\description{
The \pkg{camrprojects} package implements a set of
portable functions commonly used in R projects
from the Center for Addiction Medicine (CAM) at the
Mass General Brigham (MGB) hospital. A complete list
of the functions available in the package can be
obtained via the command
\code{ls(pos = "package:camrprojects")} after
loading in the package via a call to \code{library}.
}
