% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R17-Table1.R
\name{cbind_gt_groups}
\alias{cbind_gt_groups}
\title{Column bind two gt tables}
\usage{
cbind_gt_groups(
  gts,
  group_var,
  fct_group_col = "variable_grp",
  id_col = "Variable",
  keep_order = TRUE,
  hist_bar_color = "blue"
)
}
\arguments{
\item{gts}{A named list of gt tables. The names should be the levels of the factor
they were grouped by.}

\item{group_var}{A string. The name of the grouping factor variable.}

\item{fct_group_col}{A string. The name of the column in the gt tables
that groups factor levels. (e.g. for the factor levels
"Male" and "Female" this variable holds the value "Sex")}

\item{id_col}{A string. The column the two tables will be merged on. Usually
"Variable"}

\item{keep_order}{Logical. Indicator for preserving the order of vars in the
inputted tables. Defaults to TRUE.}

\item{hist_bar_color}{The color you want the histograms for numeric variables
to appear.}
}
\value{
A gt_tbl object that combines the tables in \code{gts}.
}
\description{
A helper function to column bind two gt tables. Used when running camr_make_table1
with groups.
}
