% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions.R
\name{check_for_missing}
\alias{check_for_missing}
\title{Checks for Missing Data}
\usage{
check_for_missing(
  x,
  codes = list(NA, "", as.Date("1970-01-01", format = "\%Y-\%m-\%d"))
)
}
\arguments{
\item{x}{A vector of values.}

\item{codes}{A list of different codes for
missing data (e.g., \code{NA}, \code{''}).}
}
\value{
A list with...
\itemize{
  \item \code{missing_values}: A logical vector indicating
    which values of \code{x} are missing;
  \item \code{x_no_missing}: All non-missing values of \code{x};
  \item \code{codes_for_missing}: A list with all missing
    value codes that were found in \code{x}. If no missing
    values were found, is \code{NULL}.
}
}
\description{
Given a list of different codes for missing
data (e.g., \code{NA}, \code{''}, etc.),
identifies missing data in a vector and
also determines which missing data codes are
applicable.
}
\details{
Vectors of class \code{Date} are
handled slightly differently, as comparisons
against values that are not dates will return
\code{NA}. Therefore, dates are only checked
against other dates and for \code{NA} values.
}
\examples{
# Vector with two types of missing values
x <- c( 'A', 'B', '', NA, 'C' )
check_for_missing( x )

# Dates
x <- as.Date( c( '2000-01-01', '1970-01-01', '2000-02-02', NA ),
              format = '\%Y-\%m-\%d' )
check_for_missing( x )

}
\author{
Kevin Potter
}
