% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R22-yaml_codebook_maker.R
\name{clean_codebook_df}
\alias{clean_codebook_df}
\title{Clean codebook dataframe}
\usage{
clean_codebook_df(df)
}
\arguments{
\item{df}{A dataframe built from yaml codebooks. Expects the row names
to be the names of variables.}
}
\value{
A dataframe with a row for each variable.
}
\description{
Takes in a codebook dataframe and processes it to a standard, clean form
with fixed set of columns.
}
