% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Wrangling.R
\name{clmn}
\alias{clmn}
\title{Extract Column Names Meeting Inclusion/Exclusion Criteria}
\usage{
clmn(dtf, ...)
}
\arguments{
\item{dtf}{A data frame.}

\item{...}{Character strings with the sub-strings to match
(or exclude) against the column names in \code{dtf}.
If an entry starts with either \code{!}, \code{~}, or
\code{-}, any columns containing the substring will be
excluded. Otherwise, the function will locate
all column names containing all inputted sub-strings.}
}
\value{
A vector of column names meeting the inclusion
and exclusion criteria.
}
\description{
A function that matches or excludes column names in a
data frame based on user-supplied sub-strings.
}
\examples{
# Create a data frame
dtf <- data.frame(
  IDS.INT.Subject = rep( 1:4, each = 2 ),
  SSS.CHR.Group = rep( c( 'A', 'A', 'B', 'B' ), each = 2 ),
  SSS.INT.Group = rep( c( 1, 1, 2, 2 ), each = 2 ),
  SSS.LGC.Group_A = rep( c( T, T, F, F ), each = 2 ),
  SSS.CHR.Time_point = rep( c( 'Pre', 'Post' ), 4 ),
  SSS.INT.Time_point = rep( 0:1, 4 ),
  OUT.DBL.Scores = rnorm( 8 )
)

#' # All variables containing 'SSS'
dtf \%>\% clmn( 'SSS' )

# All variables containing both 'SSS' and 'CHR'
dtf \%>\% clmn( 'SSS', 'CHR' )

# Variables containing 'SSS' but not 'CHR'
dtf \%>\% clmn( 'SSS', '~CHR' )

}
\author{
Kevin Potter
}
