% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{collapse_pums}
\alias{collapse_pums}
\title{Collapse census data by variable using survey weights}
\usage{
collapse_pums(pums_data, group_variable, weight_variable = PWGTP)
}
\arguments{
\item{group_variable}{Variable by which to group summary (e.g. Race or Female)}

\item{weight_variable}{Varible with population weights. Defaults to
PWGTP}

\item{pums_list_obj}{A processed PUMS dataframe. The output of \code{process_pums}.}
}
\value{
A list. Metadata (as a list), plus a dataframe with proportions of
population in each category.
}
\description{
Collapse PUMS microdata into representative summary by grouping variable
}
