% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{column_abbreviations}
\alias{column_abbreviations}
\title{Abbreviations and Labels for Column Names}
\usage{
column_abbreviations(
  column_name = "",
  type = NULL,
  custom = NULL,
  n = NA,
  separator = "."
)
}
\arguments{
\item{column_name}{A standardized variable name
(i.e., 'GGG.TTT.Var_name' or
'GGG.TTT.Optional.Var_name').}

\item{type}{The type of abbreviation whose
label is to be extracted.}

\item{custom}{An optional matrix with two
columns, 'Abbr' and 'Label', for
custom abbreviations and labels,
respectively.}

\item{n}{The number of rows to create when
generating a template for the matrix to
pass to the \code{custom} argument.}

\item{separator}{The punctuation used
to separate abbreviations, typically
a period.}
}
\value{
A character string. If only \code{n} is
  provided, a matrix with \code{n} rows with
  columns 'Abbr' and 'Label' to use as a
  template for custom abbreviations/labels.
}
\description{
Function to match labels describing
variable categories and data types to
abbreviations contained within a column
name.
}
\examples{
# List pre-defined abbreviations and labels
column_abbreviations( type = 'category' )
column_abbreviations( type = 'data type' )

# Match abbreviation for variable categories
column_abbreviations(
  'IDS.CHR.Subject', type = 'category'
)

# Match abbreviation for data type
column_abbreviations(
  'IDS.CHR.Subject', type = 'data type'
)

# Create template for custom abbreviations
# and labels
M <- column_abbreviations( n = 3 )
M[,1] <- c( 'BL', 'Y1', 'Y2' )
M[,2] <- c( 'Baseline', 'Year 1', 'Year 2' )
column_abbreviations( 'SMK.INT.BL.Quit_status', custom = M )
column_abbreviations( 'SMK.INT.Y2.Quit_status', custom = M )

}
