% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{contains_groups}
\alias{contains_groups}
\title{Determine Levels of a Grouping Factor Associated With a Variable}
\usage{
contains_groups(dtf, column_name, grouping_column, codes_for_missing)
}
\arguments{
\item{dtf}{A data frame.}

\item{column_name}{A character string for the column
name over which to determine all non-missing values.}

\item{grouping_column}{A character string for the column
name with the grouping factor.}

\item{codes_for_missing}{A character vector with
additional codes for missing data (e.g.,
\code{''}, \code{' '}, or user-defined codes).}
}
\value{
A character string.
}
\description{
Determines the available levels of a grouping factor
for all non-missing values of another variable,
to be passed into meta-data for data dictionary purposes.
}
\examples{
# Example data frame
dtf <- data.frame(
  x = rnorm( 10 ),
  # Grouping factor
  g = rep( LETTERS[1:5], each = 2 )
)
dtf$x[1:2] <- NA # Create missing values
contains_groups( dtf, 'x', 'g', '' )

}
\author{
Kevin Potter
}
