% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{create_folder_pathways}
\alias{create_folder_pathways}
\title{Create or Update List with Paths for Project Directory Hierarchy}
\usage{
create_folder_pathways(
  folder_pathways = NULL,
  add_path = NULL,
  redcap = TRUE,
  dropbox = TRUE,
  api = TRUE
)
}
\arguments{
\item{folder_pathways}{An optional existing list with paths and labels
that can then be updated with the \code{add_path} argument. If
\code{NULL} (the default), automatically generates the list AND
the standard folder hierarchy for a CAM RStudio project, where...
\describe{
\item{Project}{The project directory}
\item{+ R}{The folder for R scripts}
\item{+ Data}{}
\item{++ Data_processing}{Processing scripts to clean the raw data}
\item{++ Raw_data_files}{Files, such as downloads from REDCap, to
be cleaned and processed}
}}

\item{add_path}{An optional named list to add a new folder path and
associated labels to an existing \code{folder_pathways} object.
Must contain three elements:
\describe{
\item{path}{The absolute path to the folder to add;}
\item{title}{The internal label for the slot for the folder path;}
\item{labels}{The set of labels to use in the \code{go_to} call.}
}}

\item{redcap}{Logical; if \code{TRUE}, when creating the
\code{folder_pathways} list, will check for a folder one level
above the project directory for an installation of the R package
'REDCapR' version 0.9.8, which is compatible with current CAM
projects as of 09/04/2020.}

\item{dropbox}{Logical; if \code{TRUE}, when creating the
\code{folder_pathways} list, will check for a text file
in the 'Raw_data_files' folder that contains the path to a
dropbox folder on the local machine. If the file does not
exist, a warning will be displayed and a file template
will be created. The user can then copy the path into the
file.}

\item{api}{Logical; if \code{TRUE}when creating the
\code{folder_pathways} list, will check for a text file
in the 'Raw_data_files' folder that contains the API token
for a specified REDCap project. If the file does not
exist, a warning will be displayed and a file template
will be created. The user can then copy their API token
into the file.}
}
\value{
A list with slots for folder pathway of interest (e.g.,
the project directory, the R scripts folder, etc.), and an
additional list with labels (e.g., \code{c('Project','project')})
that can be used with the \code{\link[=go_to]{go_to()}} function to navigate
to the associated folder.
}
\description{
Creates/update a list that contains paths for the folder hierarchy
for a standard CAM RStudio project. This list can then be accessed
by the \code{\link[=go_to]{go_to()}} function for fast and readable folder
navigation.
}
\examples{
# For example purposes, create simple
# list with path and labels
folder_pathways <- list(
  Project = getwd(),
  Path_labels = list( Project = c( 'Project', 'project' ) )
)

# Add a new path and labels to the existing list
lst <- list(
  path = paste0( getwd(), '/R' ),
  title = 'R',
  labels = c( 'R', 'r' )
)
folder_pathways = create_folder_pathways( folder_pathways, add_path = lst )

}
\author{
Kevin Potter
}
