% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{create_meta_data}
\alias{create_meta_data}
\title{Create Meta-data for Data Dictionary Purposes}
\usage{
create_meta_data(
  dtf,
  column_name,
  description = "",
  variable_category = NULL,
  data_type = NULL,
  redcap_variables = "",
  values_and_labels = "",
  scale_details = NULL,
  subscale_details = NULL,
  summary_of_x = "",
  category_labels = NULL,
  type_labels = NULL,
  codes_for_missing = list(NA, "", " "),
  group_var = NULL,
  time_var = NULL,
  study_var = NULL,
  units_of_x = "",
  validated = "",
  digits = 2
)
}
\arguments{
\item{dtf}{A data frame.}

\item{column_name}{...}

\item{description}{...}

\item{variable_category}{...}

\item{data_type}{...}

\item{redcap_variables}{...}

\item{values_and_labels}{...}

\item{scale_details}{...}

\item{subscale_details}{...}

\item{summary_of_x}{...}

\item{category_labels}{...}

\item{type_labels}{...}

\item{codes_for_missing}{...}

\item{group_var}{...}

\item{study_var}{...}

\item{units_of_x}{...}

\item{validated}{...}

\item{digits}{...}
}
\value{
An updated data frame with new attributes for
the specified column.
}
\description{
Updates the attributes for a column in a data frame
with a standardized list of details on the variable
that can be then used to create a data dictionary.
Columns are assumed to be named according to a
standardized scheme, typically in the style
'AAA.BBB.Description', where 'AAA' is a 3-letter
abbreviation giving the overarching category for
a variable (e.g., 'Identifiers', 'Session details', etc.),
'BBB' is a 3-letter abbreviation for the type of data
(e.g., 'Integer', 'Character string', etc.), and
'Description' is a human-readable label
(e.g., 'Treatment_groups', 'Session_dates', etc.).
}
\author{
Kevin Potter
}
