% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{create_summary_for_x}
\alias{create_summary_for_x}
\title{Compute Brief Set of Summary Statistics for a Numeric Variable}
\usage{
create_summary_for_x(x, codes_for_missing, digits = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{codes_for_missing}{A character vector with
additional codes for missing data (e.g.,
\code{''}, \code{' '}, or user-defined codes).}

\item{digits}{Number of digits to round to.}
}
\value{
A character string with the observed summary
statistics.
}
\description{
Given a numeric variable \code{x}, computes either the mean,
median, standard deviation, and 1st and 3rd quartiles
(for continuous values) or the percentage and observed
frequencies (for binary values). Also determines the
frequency of missing data.
}
\examples{
# Create numeric vector
x <- rnorm( 100 )
# Add missing data
x[ sample( 1:100, size = 10 ) ] <- NA
# Compute summary
create_summary_for_x( x, '' )

# Binary values
x <- rbinom( 100, 1, .5 )
# Add missing data
x[ sample( 1:100, size = 10 ) ] <- NA
# Compute summary
create_summary_for_x( x, '' )

}
\author{
Kevin Potter
}
