% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{data_check_hist}
\alias{data_check_hist}
\title{Generate histogram of time between visits and initial manual data checks}
\usage{
data_check_hist(df_redcap_raw, out = NULL)
}
\arguments{
\item{df_redcap_raw}{A dataframe containing raw REDCap study data outputted by the data processing pipeline.}

\item{out}{An optional file path to save out an SVG file containing the plot.}
}
\value{
A ggplot object plotting the time between visits and manual data checks in REDCap.
}
\description{
Function that plots the time between visits and initial manual data checks.
}
