% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{data_frame_from_dictionary_meta_data}
\alias{data_frame_from_dictionary_meta_data}
\title{Create Data Frame From Dictionary Meta-Data}
\usage{
data_frame_from_dictionary_meta_data(dtf, progress = FALSE)
}
\arguments{
\item{dtf}{A data frame with columns whose
attributes include \code{dictionary_meta_data}.}

\item{progress}{Logical; if \code{TRUE}
tracks progress of extracting meta-data from
columns - useful for debugging purposes.}
}
\value{
A long-form data frame with details
(e.g., a variable's category, data type,
sub-category, description, units of measurement,
codes for missing data, scale/subscale information,
and specific notes) for each column with
dictionary meta-data.
}
\description{
Function to create a data frame with details
from the dictionary meta-data - can then
be converted into, for example, a .csv
file to easily share a data dictionary
with collaborators.
}
