% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{Datetimes_as_strings}
\alias{Datetimes_as_strings}
\alias{camr_ymd}
\alias{camr_ymd_hm}
\alias{camr_ymd_hms}
\alias{camr_hms}
\title{Functions to Extract Dates and Times as Strings}
\usage{
camr_ymd(sep = "_")

camr_ymd_hm(sep = "_", split = "-")

camr_ymd_hms(sep = "_", split = "-")

camr_hms(sep = "_")
}
\arguments{
\item{sep}{A character string, the separator between
the year, month, day, and/or hour, minutes, and seconds.}

\item{split}{A character string, the separator between
the date and the time.}
}
\value{
A character string.
}
\description{
Functions that extract the current date and time (in UTC)
and convert them to formatted strings.
}
\examples{
# YYYY_MM_DD
camr_ymd()
# YYYY-MM-DD
camr_ymd('-')
# YYYYMMDD
camr_ymd('')

# YYYY_MM_DD-HH_MM
camr_ymd_hm()
# YYYYMMDD-HHMM
camr_ymd_hm('')
# YYYYMMDD_HHMM
camr_ymd_hm('', '_')

# YYYY_MM_DD-HH_MM_SS
camr_ymd_hms()
# YYYYMMDD-HHMMSS
camr_ymd_hms('')
# YYYYMMDD_HHMMSS
camr_ymd_hm('', '_')

# HH_MM_SS
camr_hms()
# HH:MM:SS
camr_hms(':')
}
\author{
Kevin Potter
}
