% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{download_census}
\alias{download_census}
\title{Download PUMS from census data}
\usage{
download_census(
  census_api_key_path,
  cache_dir = "~/.cache/pums",
  additional_vars = NULL,
  year = 2023,
  survey = "acs1"
)
}
\arguments{
\item{census_api_key_path}{A string with the path to a txt file
with an API key for census.gov. Not required if data are cached.}

\item{cache_dir}{A string with the directory to cache the pums data
Defaults to "~/.cache/pums". Creates the directory
if it does not exist.}

\item{additional_vars}{A character vector with PUMS variable names
to download}

\item{year}{An integer. The year of PUMS data to download. Note that
the processing code is based on 2023. (A warning will print
for non-2023 years). Defaults to 2023.}

\item{survey}{A string. The survey you want to download. Defaults to
"acs1".}
}
\value{
Side effect: downloads PUMS data to \code{cache_path}.
Invisibly returns a list with the PUMS microdata and some
metadata.
}
\description{
Function that downloads PUMS data
}
