% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{download_logging}
\alias{download_logging}
\title{Download logging data from REDCap}
\usage{
download_logging(chr_rc_uri = "", chr_rc_token = "")
}
\arguments{
\item{chr_rc_uri}{A character string, the URL to use for the httr::POST function.}

\item{chr_rc_token}{A character string, the user's API token for the specified REDCap project.}
}
\value{
A dataframe containing unprocessed REDCap logging information.
}
\description{
Function that downloads logging data from a REDCap project
}
