% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{dqr_date_table}
\alias{dqr_date_table}
\title{Generate data quality rule date table}
\usage{
dqr_date_table(df_logging, users_exclude = NULL, out = NULL)
}
\arguments{
\item{df_logging}{A dataframe containing processed REDCap logging information.}

\item{users_exclude}{An optional vector of REDCap usernames listing who to exclude from data quality rule execution.}

\item{out}{An optional file path to save out an HTML file containing the table.}
}
\value{
A gt table printing the last 3 dates the data quality rules were executed.
}
\description{
Function that pulls the last 3 dates the data quality rules were run. Expects dataframe output from \code{process_logging}.
}
