% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S09-elements.R
\name{elements}
\alias{elements}
\title{Extract Unique Elements of a Vector}
\usage{
elements(x, counts = T, include = NULL, exclude = "", na.rm = T)
}
\arguments{
\item{x}{A vector.}

\item{counts}{Logical; if \code{TRUE} returns a count of the number
of unique values, otherwise returns a vector with the unique values.}

\item{include}{Optional logical vector of matching length to \code{x}
indicating obserations to keep.}

\item{exclude}{Optional vector of types of observations to exclude
(exact matches).}

\item{na.rm}{Logical; if \code{TRUE}, removes \code{NA} values.}
}
\value{
A count of the number of unique values or a vector
of the unique values.
}
\description{
A convenience function that extracts the unique elements
of a vector, after optionally excluding/keeping specified
observations. The function can summarize the number of
unique values, or return the unique values as a vector.
}
\examples{
vec <- c( rep( c( 'A', 'B', 'C', 'D' ), each = 3 ), NA )
# Number of unique elements - NA excluded by default
elements( vec )
# Unique elments as vector
elements( vec, count = F )
# Keep NA
elements( vec, na.rm = F )
elements( vec, na.rm = F, count = F )
# Exclude specific values
elements( vec, exclude = c( 'A', 'B' ) )
elements( vec, exclude = c( 'A', 'B' ), count = F )
elements( vec, exclude = c( 'A', 'B' ), count = F, na.rm = F )

}
\author{
Kevin Potter
}
