% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R22-yaml_codebook_maker.R
\name{extract_yaml}
\alias{extract_yaml}
\title{Extracted needed info from yaml}
\usage{
extract_yaml(varname, yaml_file)
}
\arguments{
\item{varname}{String. Name of target variable.}

\item{yaml_file}{String. Path to yaml file containing info on \code{varname}}
}
\value{
List. List of info on target variable from yaml codebook file.
}
\description{
Grabs variable and form info, for a target variable, from a yaml codebook file
}
