% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04-file_present.R
\name{file_present}
\alias{file_present}
\title{Checks if a File is Present in a Folder}
\usage{
file_present(string, output = "Logical")
}
\arguments{
\item{string}{A file name or part of a file name to search for.}

\item{output}{Character string indicating the type of output to
  return, either...
\enumerate{
  \item 'Logical' or 'logical'
  \item 'Vector', 'vector', or 'vec'
  \item 'Index' or 'index'
  \item 'Name' or 'name'
}}
}
\value{
Either...
\enumerate{
  \item A logical value, \code{TRUE} if the file is present;
  \item A logical vector for all files in the folder;
  \item The index position for the file if it exists;
  \item The file name.
}
}
\description{
Checks if a file is present in the current working directory.
}
\author{
Kevin Potter
}
