% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{find_range_for_x}
\alias{find_range_for_x}
\title{Determine the Observed Range for a Numeric Variable}
\usage{
find_range_for_x(x, codes_for_missing, digits = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{codes_for_missing}{A character vector with
additional codes for missing data (e.g.,
\code{''}, \code{' '}, or user-defined codes).}

\item{digits}{Number of digits to round to.}
}
\value{
A character string, formatted as
\code{'Observed = X to Y'} where 'X' is the
minimum observed value and 'Y' is the maximum
observed value after removing missing data.
}
\description{
Given a numeric vector \code{x}, determines the
observed minimum and maximum after removing
missing data and outputs a character string to
be passed into the meta-data for data dictionary
purposes.
}
\examples{
# Create numeric vector
x <- rnorm( 100 )
# Add missing data
x[ sample( 1:100, size = 10 ) ] <- NA
# Determine range
find_range_for_x( x, '' )

}
\author{
Kevin Potter
}
