% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R15-data_quality_rules_gen.R
\name{gen_data_quality_rules}
\alias{gen_data_quality_rules}
\title{Generate data quality rules based on REDCap metadata}
\usage{
gen_data_quality_rules(
  path_to_api_token,
  output_path = "data_quality_rules.csv"
)
}
\arguments{
\item{path_to_api_token}{path to a text file containing project API token}

\item{output_path}{path to a file where csv should be saved; will \emph{not}
overwrite an existing file. Defaults to "data_quality_rules.csv"}
}
\value{
side-effect: saves csv at \code{output_path}. Invisibly returns the
data quality rules csv for inspection
}
\description{
Write data quality rules based on REDCap Metadata. Values should not be
missing when they are 1) Required AND 2) reached by branching logic.
Produces a csv that can be uploaded to the data quality module in REDCap.f
}
