% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{get_pums_summaries}
\alias{get_pums_summaries}
\title{Function to get summary dataframe of PUMS data}
\usage{
get_pums_summaries(pums_data)
}
\description{
Loop over the grouping variables in processed PUMS data and produce a
list of summary dataframes
}
