% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R99-Legacy.R
\name{go_to}
\alias{go_to}
\title{Navigate to Folder in Project Directory}
\usage{
go_to(path)
}
\arguments{
\item{path}{The label associated with the desired folder to navigate to
(e.g., 'Project', 'R', etc.).}
}
\description{
A convenience function that allows a user to quickly navigate
to specified folders in the Project directory via an easily
readable command. Requires a list called \code{folder_pathways}
with the paths and associated labels to exist in the global environment
(see \code{\link[=create_folder_pathways]{create_folder_pathways()}}).
}
\author{
Kevin Potter
}
