% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{\%em\%}
\alias{\%em\%}
\title{Binary Operator for Exact Matches}
\usage{
x \%em\% y
}
\arguments{
\item{x}{A vector of values to be matched.}

\item{y}{The value to be exactly matched against.}
}
\value{
A logical vector equal to \code{TRUE} for all elements
in \code{x} that exactly match \code{y}.
}
\description{
The binary operator \code{\%em\%} returns a logical
vector indicating if there is a exact match against its left
operand.
}
\details{
This operator is robust against \code{NA}, returning
\code{TRUE} if \code{y} equals \code{NA} and \code{FALSE}
otherwise.
}
\examples{
x <- c( "Cat", "Dog", "Bat" )
x \%em\% "Cat"

# Can match NA
x[2] <- NA
x \%em\% "Cat"
x \%em\% NA

}
\author{
Kevin Potter
}
