% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Binary Operator for Non-Matches}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector of values to be matched.}

\item{y}{A vector of values to be matched against.}
}
\value{
A logical vector equal to \code{TRUE} for all elements
in \code{x} that did not match any elements in \code{y}.
}
\description{
The binary operator \code{\%not_in\%} returns a logical
vector indicating if there is no matches for its left
operand.
}
\examples{
x <- c( "Cat", "Dog", "Bat" )
y <- c( "Cat", "Dog" )

x \%not_in\% y

}
\author{
Kevin Potter
}
