% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{\%p\%}
\alias{\%p\%}
\title{Operator to Concatenate Two Strings}
\usage{
x \%p\% y
}
\arguments{
\item{x, y}{\strong{R} objects that can be converted
to character vectors.}
}
\value{
A character vector of the concatenated values.
}
\description{
The operator \code{\%p\%} combines character vectors.
}
\details{
The call \code{ x \%p\% y } is equivalent to the
call \code{ paste0(x,y)}; see \code{\link[base]{paste}}
for more details.
}
\examples{
# Combine strings via operator
"Hello" \%p\% " " \%p\% "world"

# Vectorized
x <- "I like "
y <- c("cats", "dogs", "fish")
x \%p\% y

}
\author{
Kevin Potter
}
