% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R00-Utilities_and_operators.R
\name{\%pm\%}
\alias{\%pm\%}
\title{Binary Operator for Partial Matches}
\usage{
x \%pm\% y
}
\arguments{
\item{x}{A vector of values to be matched.}

\item{y}{The value to be partially matched against.}
}
\value{
A logical vector equal to \code{TRUE} for all elements
in \code{x} that contain \code{y} in some form.
}
\description{
The binary operator \code{\%pm\%} returns a logical
vector indicating if there is a partial match against its left
operand (see \link[base:grep]{base::grepl}).
}
\details{
This operator is robust against \code{NA}, returning
\code{TRUE} if \code{y} equals \code{NA} and \code{FALSE}
otherwise.
}
\examples{
x <- c( "Cat", "Dog", "Bat" )
x \%pm\% "at"

# Can match NA
x[2] <- NA
x \%pm\% "at"
x \%pm\% NA

}
\author{
Kevin Potter
}
