% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R13-Initialize_Pipeline.R
\name{init_pipeline}
\alias{init_pipeline}
\title{Initialize a new pipeline based on a REDCap project and the standard pipeline}
\usage{
init_pipeline(
  token_file,
  project_name,
  nickname,
  dest_dir,
  std_repo_url,
  std_repo_branch = "main",
  redcap_url = Sys.getenv("REDCAP_API_URL")
)
}
\arguments{
\item{token_file}{Path to a text file that contains \strong{only} the API token
for the target REDCap project.}

\item{project_name}{A descriptive project title (usually matching the REDCap
project name, but not necessarily)}

\item{nickname}{Short prefix to apply to all project-specific functions}

\item{dest_dir}{Directory in which the new project should be created. Must
\emph{not} exist (it will be created by this function)}

\item{std_repo_url}{HTTPS or SSH URL of the standard pipeline repository}

\item{std_repo_branch}{Branch to clone from std_repo_url (default "main")}

\item{redcap_url}{API endpoint. Defaults to Sys.getenv("REDCAP_API_URL") and
aborts if still empty}
}
\value{
side effect: \code{dest_dir} is populated with a provisional targets pipeline.
\code{dest_dir} is also invisibly returned as a string.
}
\description{
Clone the standard pipeline, pull REDCap project metadata, and
draft an R project directory with _targets.R file stubs to match instruments
in the REDCap project. For instruments with processing functions in the
standard pipeline, the corresponding processing functions are copied and
renamed with the prefixes based on the provided project nickname. Instruments
without standard processing functions will have placeholder functions and targets.
}
\examples{
init_pipeline(
  token_file      = "~/Documents/api_tokens/groundbreakingstudy.txt",
  project_name.   = "Groundbreaking Substance Use Study"
  nickname        = "gbsus"
  dest_dir.       = "~/Documents/code_repos/gbsus_2025"
  std_repo_url    = "git@gitlab.ourOrg.org:our_team/standard-data-pipeline.git"
  redcap_url.     = "https://redcap.ourOrg.org/redcap/api/"
)
}
