% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R07-Codebook.R
\name{is.codebook_entry}
\alias{is.codebook_entry}
\title{Check if a Variable is of Class Codebook Entry}
\usage{
is.codebook_entry(x)
}
\arguments{
\item{x}{An R object to be checked.}
}
\value{
A logical value, \code{TRUE} if \code{x}
is of class \code{codebook_entry}.
}
\description{
Method to check if an R object is of class
\code{codebook_entry}.
}
\author{
Kevin Potter
}
