% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{is.dictionary_meta_data}
\alias{is.dictionary_meta_data}
\title{Check if a Variable is of Class Dictionary Meta-data}
\usage{
is.dictionary_meta_data(x)
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
