% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{is.meta_data_for_variable}
\alias{is.meta_data_for_variable}
\title{Title}
\usage{
is.meta_data_for_variable(x)
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
