% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{known_scales}
\alias{known_scales}
\title{Pre-defined Output for Known Scales and Subscales}
\usage{
known_scales(abbreviation = NULL, subscale = "")
}
\arguments{
\item{abbreviation}{The abbreviation for a scale, used to
look up the associated pre-defined output.}

\item{subscale}{The name of a subscale of the scale
given above, used to look up the relevant ancillary
output.}
}
\value{
Either 1) a list with the scale name, number of items,
vector with the lowest and largest possible values,
abbreviation, a vector of any clinical cut-off values, and
a character vector with APA-style references for the scale,
or 2) a list with the subscale name, number of items, vector
with the lowest and largest possible values, and a vector
with any clinical cut-off values.
}
\description{
A convenience function that, given an abbreviation
for a scale, inventory, or questionnaire, returns
a list of information useful for data dictionary
purposes.
}
\examples{
# List all possible inputs (i.e., known
# scales, inventories, and questionnaires)
known_scales()

# Example input
known_scales( 'HADS' )
known_scales( 'HADS', subscale = 'Anxiety' )

}
