% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{label_for_abbr}
\alias{label_for_abbr}
\title{Extract Labels Corresponding to Abbreviations}
\usage{
label_for_abbr(column_name, type = "category", abbr_labels = NULL, pst = NULL)
}
\arguments{
\item{column_name}{The column name in a data frame,
assumed to follow a standardized format, such
as 'AAA.BBB.Description' or 'AAA.BBB.CCC.Description'.}

\item{type}{A character string indicating the type of
abbreviation/label pairings, either...
\itemize{
  \item \code{'category'} for abbreviations associated
  with a variable's overarching category (e.g., 'Identifier',
  'Session details', etc.);
  \item \code{'data type'} for abbreviations associated
  with the type of data (e.g., 'Integer', 'Character string',
  etc.);
  \item \code{'custom'} for custom abbreviations and labels
  for extensions beyond the typical 'AAA.BBB.Description'
  column name construction.
}}

\item{abbr_labels}{A 2-column character matrix, with the first
column listing abbreviations and the second column listing
associated labels. Can be used to provide additional
abbreviation/label pairings in addition to the default
options for variable categories and data types, or can
be used in conjunction with the argument \code{pst} to
specify completely custom abbreviations and labels.}

\item{pst}{A vector of two integers giving the starting and
ending positions in the column name character string with
the abbreviation to extract.}
}
\value{
A character string, the corresponding label
  to the abbreviation extracted from \code{column_name}.
}
\description{
Given a standardized naming schemes for columns in
a data frame (e.g., \code{AAA.BBB.Description} where
'AAA' is a 3-letter abbreviation for the overarching
category to which a variable belongs, and 'BBB' is a
3-letter abbreviation for the type of data in the column,
this function extracts the corresponding label associated
with a given abbreviation. Additional abbreviations and
associated labels can be defined for the categories/data types,
and a completely custom set of abbreviations (e.g., 2-letter
abbreviations for time points) and associated labels
can be defined.
}
\examples{
# Determine variable's category
label_for_abbr( 'SSS.CHR.Subject' )

# Determine variable's data type
label_for_abbr( 'SSS.CHR.Subject', type = 'data type' )

# Add additional labels
al <- rbind(
  c( 'MED', 'Medication use' ),
  c( 'DRV', 'Driving simulation results' )
)
label_for_abbr( 'MED.INT.Bupropion_use', abbr_labels = al )
label_for_abbr( 'DRV.INT.Divided_attention_task', abbr_labels = al )

# Custom abbreviations and labels
al <- rbind(
  c( 'BL', 'Baseline' ),
  c( 'Y1', 'Year 1 of intervention' ),
  c( 'Y2', 'Year 2 of intervention' )
)
pst <- c( 9, 10 )
label_for_abbr( 'SMK.INT.BL.Smoking_status',
                type = 'custom', abbr_label = al, pst = pst )
label_for_abbr( 'SMK.INT.Y2.Smoking_status',
                type = 'custom', abbr_label = al, pst = pst )

}
\author{
Kevin Potter
}
