% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-StatisticalTools.R
\name{limits_for_interp}
\alias{limits_for_interp}
\title{Function to Find Limits for Interpolation}
\usage{
limits_for_interp(
  value,
  x,
  y,
  interp_y = T,
  use_first_last = F,
  warn_if_first_last = T
)
}
\arguments{
\item{value}{The x or y-axis value to interpolate.}

\item{x}{A vector of numeric values for the x-axis
(must match \code{y} in length).}

\item{y}{A vector of numeric values for the y-axis
(must match \code{x} in length).}

\item{interp_y}{Logical; if \code{TRUE}, the value to
interpolate is assumed to be a y coordinate.}

\item{use_first_last}{Logical; if \code{TRUE} when values
fall outside the provided vectors the function uses the
first or last set of x and y values as an approximation.}

\item{warn_if_first_last}{Logical; if \code{TRUE} warns
the user when an approximation is used for cases that
fall outside the range of the provided vectors.}
}
\value{
A vector of 5 values, the lower x and y-axis values
followed by the upper x and y-axis values that bracket the
point to interpolate. If there are no coordinates bracketting
the specified coordinate, a vector of \code{NA} values is returned.
}
\description{
A function that, given a vector of x and y values,
identifies the lower and upper bounds to use for
linear interpolation.
}
\examples{
# Example curve
x = 0:5
y = c( 0, .5, 2, 4, 8, 16 )

# Value between x-axis values 2 and 3
inp = limits_for_interp( 2.5, x, y )
linear_interp( inp )

#' # Interpolate x-axis value
inp = limits_for_interp( 3, x, y, interp_y = F )
linear_interp( inp, interp_y = F )

# Value at an existing point
inp = limits_for_interp( 3, x, y )
linear_interp( inp )

# Value out of range
inp = limits_for_interp( 6, x, y )
linear_interp( inp )

# Approximation with first/last value (use with caution)
inp = limits_for_interp( 6, x, y, use_first_last = T )
linear_interp( inp )

}
\author{
Kevin Potter
}
