% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-StatisticalTools.R
\name{linear_interp}
\alias{linear_interp}
\title{Function for Linear Interpolation}
\usage{
linear_interp(vec, interp_y = T)
}
\arguments{
\item{vec}{A vector of 5 values, consisting of...
\describe{
\item{x0}{The x coordinate for the lower boundary;}
\item{y0}{The y coordinate for the lower boundary;}
\item{x1}{The x coordinate for the upper boundary;}
\item{y1}{The y coordinate for the upper boundary;}
\item{x or y}{The coordinate at which interpolation should occur.}
}
A named vector can be provided, otherwise the order is assumed to
follow the list given above.}

\item{interp_y}{Logical; if \code{TRUE}, the value to
interpolate is assumed to be a y coordinate.}
}
\value{
The interpolated x or y value. If outside the lower or upper
boundaries, \code{NA} is returned instead.
}
\description{
Given a pair of x and y values, uses linear
interpolation to compute a new x or y value.
}
\examples{
# Linear interpolation for y
linear_interp( c( x0 = 0, y0 = 0, x1 = 1, y1 = 2, x = .5 ) )
# Linear interpolation for x
linear_interp( c( x0 = 0, y0 = 0, x1 = 1, y1 = 2, y = .5 ), FALSE )
# Linear interpolation across multiple values
x = c( 0, 1, 2 )
y = c( 0, 2, 4 )
# Create matrix with 5th column for points to interpolate at
m = cbind( x[-3], y[-3], x[-1], y[-1], c( .5, 1.5 ) )
# Linear interpolation for y values
apply( m, 1, linear_interp )
# Linear interpolation for x values (NA for value outside boundary)
apply( m, 1, linear_interp, interp_y = F )

}
\author{
Kevin Potter
}
