% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{list_of_missing}
\alias{list_of_missing}
\title{Creates List of Missing Data Found in a Vector}
\usage{
list_of_missing(x, codes_for_missing = c("", " "))
}
\arguments{
\item{x}{A vector of values.}

\item{codes_for_missing}{A character vector with
additional codes for missing data (e.g.,
\code{''}, \code{' '}, or user-defined codes).}
}
\value{
A character string listing the types of missing data
found. For example, if a vector uses R's standard \code{NA}
and \code{NaN} values to indicate missing values, the
output would be: \code{'<NA|NaN>'}. Different codes for
missing values are separated by the pipe symbol \code{|}.
}
\description{
Creates a formatted character string listing the
different types of missing data found in a vector
to then pass into the meta-data used for data
dictionary purposes.
}
\examples{
# Numeric example
x <- rnorm( 10 ); x[1] <- NA
list_of_missing( x )

# Character vector with '' and ' ' indicating missingness
x <- LETTERS; x[1] <- ''; x[2] <- ' '
list_of_missing( x )

}
\author{
Kevin Potter
}
