% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-load_package.R
\name{load_package}
\alias{load_package}
\title{Installs and Loads an R Package}
\usage{
load_package(package_name, from = "CRAN", repo = NULL, ...)
}
\arguments{
\item{package_name}{A character string with the package name.}

\item{from}{Where to download the package from; Options include...
#' \enumerate{
  \item 'CRAN'
  \item 'Github'
}}

\item{repo}{An optional character string with the Github repository
name (in the form 'username/repository') If \code{NULL}, assumes the username is
'rettopnivek' and that the repository is 'package_name'.}

\item{...}{Additional parameters for the 'devtools' installation functions.}
}
\description{
Checks if a package is installed, and if not, sees if it
can be installed. Then, loads the package for easy use.
}
\examples{
# Install/load package from CRAN
load_package( 'dplyr' )

# Install/load package from author's github repository
load_package( 'utilityf', from = 'Github' )

}
\author{
Kevin Potter
}
