% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R11-Datasets.R
\docType{data}
\name{lst_asri_scoring}
\alias{lst_asri_scoring}
\title{ASRI Scoring Tables}
\format{
A list with 3 dataframes:
\describe{
\item{probability}{Raw scores and associated probability}
\item{raw}{Raw score calculated for each sex, age, question, and response}
\item{tscore}{T-score corresponding to each sex, age, and raw score}
}
}
\usage{
data(lst_asri_scoring)
}
\description{
A list containing three dataframes with scoring tables
for the ASRI (Conners ADHD Self-Report Index).
}
\examples{

data(lst_asri_scoring)

}
\keyword{datasets}
