% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R11-Datasets.R
\docType{data}
\name{lst_meddra}
\alias{lst_meddra}
\title{MedDRA Database}
\format{
A list with 3 dataframes:
\describe{
\item{codes}{All MedDRA codes and the corresponding level}
\item{codes$levels}{MedDRA level}
\item{codes$code}{MedDRA code}
\item{terms}{All level MedDRA terms and codes}
\item{terms$llt_code}{MedDRA code to identify the Lowest Level Term}
\item{terms$pt_code}{MedDRA code to identify the Preferred Term}
\item{terms$hlt_code}{MedDRA code to identify the High Level Term}
\item{terms$hlgt_code}{MedDRA code to identify the High Level Group Term}
\item{terms$soc_code}{MedDRA code to identify the System Organ Class}
\item{terms$llt_name}{Full name of the Lowest Level Term}
\item{terms$pt_name}{Full name of the Preferred Term}
\item{terms$hlt_name}{Full name of the High Level Term}
\item{terms$hlgt_name}{Full name of the High Level Group Term}
\item{terms$soc_name}{Full name of the System Organ Class}
\item{terms$soc_abbrev}{System Organ Class abbreviation}
\item{terms$pt_soc_code}{The primary System Organ Class to which the Preferred Term is linked}
\item{terms$primary_soc_fg}{Flag to indicate primary System Organ Class}
\item{codebook}{Data codebook for MedDRA terms}
\item{codebook$varname}{Variable name for terms dataframe}
\item{codebook$datatype}{Variable data type for terms dataframe}
\item{codebook$description}{Variable description for terms dataframe}
}
}
\usage{
data(lst_meddra)
}
\description{
A list containing three dataframes of the MedDRA database.
Allows individuals to take downloaded MedDRA codes from
REDCap and pull corresponding terms and MedDRA categories.
}
\examples{

data(lst_meddra)

}
\keyword{datasets}
