% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R11-Datasets.R
\docType{data}
\name{lst_promis29_scoring}
\alias{lst_promis29_scoring}
\title{PROMIS-29 Scoring Tables}
\format{
A list with 7 dataframes:
\describe{
\item{ability_to_participate}{Raw scores, t-scores and SEs for the Ability
to Participate in Social Roles and Activities domain}
\item{anxiety}{Raw scores, t-scores and SEs for the Anxiety domain}
\item{depression}{Raw scores, t-scores and SEs for the Depression domain}
\item{fatigue}{Raw scores, t-scores and SEs for the Fatigue domain}
\item{pain_interference}{Raw scores, t-scores and SEs for the Pain
Interference domain}
\item{physical_function}{Raw scores, t-scores and SEs for the Physical
Function domain}
\item{sleep_disturbance}{Raw scores, t-scores and SEs for the Sleep
Disturbance domain}
}
}
\usage{
data(lst_promis29_scoring)
}
\description{
A list containing seven dataframes with scoring tables
for the PROMIS-29 subscores.
}
\examples{

data(lst_promis29_scoring)

}
\keyword{datasets}
