% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R11-Datasets.R
\docType{data}
\name{lst_wais_scoring}
\alias{lst_wais_scoring}
\title{WAIS-IV Digit Span Scoring Tables}
\format{
A list with 4 dataframes:
\describe{
\item{dsb}{Age and raw and scaled scores for the Digit Span Backward}
\item{dsf}{Age and raw and scaled scores for the Digit Span Forward}
\item{dss}{Age and raw and scaled scores for the Digit Span Sequencing}
\item{total}{Age and raw and scaled scores for the Digit Span Total}
}
}
\usage{
data(lst_wais_scoring)
}
\description{
A list containing four dataframes with scoring tables
for the WAIS-IV Digit Span.
}
\examples{

data(lst_wais_scoring)

}
\keyword{datasets}
