% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R11-Datasets.R
\docType{data}
\name{lst_wtar_scoring}
\alias{lst_wtar_scoring}
\title{WTAR Scoring Tables}
\format{
A list with 3 dataframes:
\describe{
\item{fsiq}{Standard score and corresponding VIQ, PIQ, and FSIQ}
\item{percentile}{Standard score and corresponding percentile}
\item{standard}{Age, raw score, and corresponding standard score}
}
}
\usage{
data(lst_wtar_scoring)
}
\description{
A list containing three dataframes with scoring tables
for the WTAR (Weschler Test of Adult Reading).
}
\examples{

data(lst_wtar_scoring)

}
\keyword{datasets}
