% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R23-REDCap_var_processing.R
\name{make_indicators}
\alias{make_indicators}
\title{Helper function to process and rename checkbox vars}
\usage{
make_indicators(
  df,
  var_name,
  answer_choices,
  prefix,
  data_type = "LGL",
  sep = "\\\\|"
)
}
\arguments{
\item{df}{A dataframe.}

\item{var_name}{A string. The name of the variable in \code{df} to be processed.}

\item{answer_choices}{A list. The possible checkboxes.}

\item{prefix}{A string. The first part of the output variable name. Usually
this will be "INV" but sometimes other choices will be used (e.g. "SBJ" or "IDX")}

\item{data_type}{A string. The data type to be used in renaming. Defaults to "LGL".}

\item{sep}{A string. The character that separates the checkbox choices in the
raw REDCap data. Defaults to "\\|" as the data come in e.g. "1|3|5|6"}
}
\description{
Take in a df and process a single checklist REDCap variable. The variable is
split into a set of indicator (logical) columns, one for each checkbox.
}
