% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{make_pums_comparison_table}
\alias{make_pums_comparison_table}
\title{Make table comparing PUMS-derived population with sample demographics}
\usage{
make_pums_comparison_table(sample_summary_dfs, pums_summary_dfs, output_path)
}
\description{
Creates a table comparing the sex, ethnicity, and age distributions of
a sample with a population derived from the Census's PUMS data.
}
