% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S17-match_and_assign.R
\name{match_and_assign}
\alias{match_and_assign}
\title{Assign New Values Based on Partial or Exact Matching}
\usage{
match_and_assign(x, matches, new_values, type = "partial", default = NA)
}
\arguments{
\item{x}{A vector of values to match over.}

\item{matches}{A list of values in \code{x} to match over}

\item{new_values}{A vector of new values to assign based on
matches to elements from \code{matches} (vector must be
of equivalent length to \code{matches}).}

\item{type}{The type of matching, either 'partial' or 'exact'
(uses \code{grepl} or \code{\%in\%}, respectively).}

\item{default}{The default value to assign in the absence of
a match.}
}
\value{
A new vector of equivalent length to \code{x}, with
  values assigned based on successful matches.
}
\description{
Assigns new values based on partial or exact matches with
values from an input vector.
}
\examples{
x <- c( 'Cat', 'Hat', 'Rat', 'Dog', 'Fog', 'Cog' )
# Partial matching
match_and_assign( x, list( 'at', 'og' ), c('A','B') )
# Exact matching
match_and_assign( x, list( 'Cat', c( 'Dog', 'Fog' ) ), c('A','B'),
                  default = '' )

}
\author{
Kevin Potter
}
