% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{meta}
\alias{meta}
\title{Extract Meta-data for Data Frame}
\usage{
meta(dtf, column_name, element = "description", parts = FALSE)
}
\arguments{
\item{dtf}{A data frame.}

\item{column_name}{A character string for the name of
the column with the meta-data to extract.}

\item{element}{The attribute in the meta-data list
to extract. Use \code{element = 'all'} to list
all attributes.}

\item{parts}{Logical; if \code{TRUE} splits
multi-part elements (i.e., lists with elements
separated by the pipe symbol '|').}
}
\value{
A character vector.
}
\description{
Extracts a specified attribute in the list of meta-data
for data dictionary purposes.
}
\author{
Kevin Potter
}
