% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{new_dictionary_meta_data}
\alias{new_dictionary_meta_data}
\title{Constructor Function for Dictionary Meta-data Class}
\usage{
new_dictionary_meta_data(
  Column_name = "",
  Variable_category = "",
  Data_type = "",
  Sub_category = "",
  Description = "",
  REDCap_variables = "",
  Values_and_labels = "",
  Scale = "",
  Subscale = "",
  Summary = "",
  Units = "",
  Codes_for_missing = "",
  Groups_collected_over = "",
  Times_collected_over = "",
  Studies_collected_over = "",
  Validated = "",
  Notes = ""
)
}
\arguments{
\item{Column_name}{...}

\item{Variable_category}{...}

\item{Data_type}{...}

\item{Description}{...}

\item{REDCap_variables}{...}

\item{Values_and_labels}{...}

\item{Scale}{...}

\item{Subscale}{...}

\item{Summary}{...}

\item{Units}{...}

\item{Codes_for_missing}{...}

\item{Groups_collected_over}{...}

\item{Times_collected_over}{...}

\item{Studies_collected_over}{...}

\item{Validated}{...}

\item{Notes}{...}
}
\value{
A named list of class \code{dictionary_meta_data}.
}
\description{
Description.
}
\details{

}
