% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S19-notes_for_custom_function.R
\name{notes_for_custom_function}
\alias{notes_for_custom_function}
\title{Display Documention for Custom R Function}
\usage{
notes_for_custom_function(f)
}
\arguments{
\item{f}{An R function saved in the current workspace.}
}
\description{
Extracts and displays documention for a custom R function,
given a documentation format of
}
\details{
Documentation start with the phrase '  # Purpose:'.
Each subsequent line that starts with '  #' will be displayed.

Standard CAM procedure for documenting functions is to have
lines following the headers 'Purpose:', 'Arguments:', and
'Returns:'.
}
\examples{
# Example custom function with standardized documentation
f = function( x ) {
  # Purpose:
  # Displays the values of 'x'
  # Arguments:
  # x - An R object
  # Returns:
  # Output of the 'print' call

  # Internal notes
  print( x )
}
notes_for_custom_function( f )

}
\author{
Kevin Potter
}
