% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R14-Yaml_codebook.R
\name{parse_transmute_pairs}
\alias{parse_transmute_pairs}
\title{Extract pairs of new and source variables from every dplyr::transmute call in a script}
\usage{
parse_transmute_pairs(file, redcap_vars, keep = NULL)
}
\arguments{
\item{file}{Path to an R script that defines target function}

\item{redcap_vars}{a character vector of the variable names in the redcap project}

\item{keep}{Optional character vector of variable names to keep}
}
\value{
A list of lists, each with the processed variable (as key) and
source variable (as value). The value defaults to "composite"
}
\description{
Extract pairs of new and source variables from every dplyr::transmute call in a script
}
