% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R99-Legacy.R
\name{path_by_os}
\alias{path_by_os}
\title{Navigate to Folder based on OS}
\usage{
path_by_os(paths)
}
\arguments{
\item{paths}{A named vector consisting of...
\describe{
\item{Windows}{The path for Windows-based operating systems}
\item{Mac}{The path for Mac-based operating systems}
\item{Linux}{The path for Linux-based operating systems}
}}
}
\value{
A logical value, \code{TRUE} if the path
was successfully accessed.
}
\description{
A function to navigate to different folders/paths
conditional on a user's operating system.
}
\examples{
accessed_path <- path_by_os(
  c( Windows = 'C:/Users',
  Mac = '~/Users',
  Linux = '~' )
)
print( accessed_path ); print( getwd() )

}
\author{
Kevin Potter and William Schmitt
}
