% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R20-Visit_and_dropout_flow.R
\name{preprocess_visit_flow_data}
\alias{preprocess_visit_flow_data}
\title{Pre-processing of data for visit/dropout flow visualizations}
\usage{
preprocess_visit_flow_data(
  df_visits,
  df_status,
  dropout_statuses = c("LTFU", "Withdrawn", "Terminated")
)
}
\arguments{
\item{df_status}{A dataframe following the conventions for \code{df_status} in the
standard pipeline. Expects dropouts to have \code{SBJ.FCT.Status} of
"LTFU", "Withdrawn", or "Terminated".}

\item{df_visit}{A dataframe following the conventions for \code{df_visits} in the
standard pipeline. The function expects SSS.LGL.Visit.Attd to indicate
visit completion, IDX.INT.VisitNumber to provide the ordered visit numbers, and
SSS.FCT.VisitName to provide visit names (with levels that match the visit numbers).}
}
\value{
A list of processed dataframes for vizualization functions.
}
\description{
Takes in visit and status dataframes and prepares them for visualization
}
