% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{print.dictionary_meta_data}
\alias{print.dictionary_meta_data}
\title{Title}
\usage{
\method{print}{dictionary_meta_data}(x, digits = 2)
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
