% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-Functions_for_REDCap_forms.R
\name{process_comm}
\alias{process_comm}
\title{Process data from the COMM questionnaire}
\usage{
process_comm(comm_dtf)
}
\arguments{
\item{rc_dtf}{The unprocessed, direct from REDCap data frame. Required.}

\item{dtf}{The processed data frame that you can assume has the
necessary session details. Required.}
}
\value{
\code{dtf} with new columns that have meta-data (i.e. attributes)
filled out.
}
\description{
This function scores the COMM questionnaire by summing the individual item
ratings and providing the subsequent indication. The function takes in the
processed data frame of REDCap data, subsets just its COMM variables, scores
the COMM (creating a sum and indication column), then adds these columns
back to the larger processing data frame.
}
\author{
Eylul Akman
}
