% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{process_logging}
\alias{process_logging}
\title{Process logging data from REDCap}
\usage{
process_logging(
  df_sl_subject_ids,
  df_vl_visits,
  df_logging_raw,
  record_col = IDS.CHR.Record,
  visit_date_col = SSS.DAT.Visit
)
}
\arguments{
\item{df_sl_subject_ids}{A dataframe from the data processing pipeline containing record and subject identifiers.}

\item{df_vl_visits}{A dataframe from the data processing pipeline containing visit dates.}

\item{df_logging_raw}{A dataframe containing raw logging data from REDCap.}

\item{record_col}{A column name specifying the record IDs in df_sl_subject_ids.}

\item{visit_date_col}{A column name specifying the visit dates in df_vl_visits.}
}
\value{
A dataframe containing processed REDCap logging information.
}
\description{
Function that processes raw logging data from REDCap. Expects dataframe output from \code{download_logging}.
}
