% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{process_pums}
\alias{process_pums}
\title{Process PUMS Census Data (for comparison with CAM study samples)}
\usage{
process_pums(
  pums_list_obj,
  age_lower_limit = NULL,
  age_upper_limit = NULL,
  states_to_drop = NULL,
  states_to_keep = NULL
)
}
\arguments{
\item{age_lower_limit}{Optional. An integer. The minimum age (inclusive) for your
population of interest.}

\item{age_upper_limit}{Optional An integer. The maximum age (inclusive) for your
population of interest.}

\item{states_to_drop}{Optional. A character vector. The abbreviations of the states
NOT in your population (e.g. c("PR", "DC")).}

\item{states_to_keep}{Optional. A character vector. The abbreviations of the
states (e.g. c("NY", "MA")) that delimit your population.}

\item{pums_list}{A list. The list output by \code{download_census}, where the
first element is metadata and the second element is a
dataframe.}
}
\value{
A list with a metadata element and the dataframe containing
the processed PUMS data
}
\description{
Process raw PUMS data to prepare for comparison with CAM study data. Expects
list output from \code{download_census}. By default the processing retains
repondents of all ages from all 50 states (plus DC and Puerto Rico.)
}
