% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R23-REDCap_var_processing.R
\name{process_redcap_var}
\alias{process_redcap_var}
\title{Helper function to process individual variables}
\usage{
process_redcap_var(df, var_name, redcap_var_metadata, prefix)
}
\arguments{
\item{df}{A dataframe.}

\item{var_name}{A string. The name of the variable to be processed.}

\item{redcap_var_metadata}{A dataframe. The output of
\code{camrprojects::camr_redcap_field_meta(api_token_path)}. Contains info about
variable types and text field validation. Used to determine how a variable
should be processed (e.g. dropdowns are converted to factors)}

\item{prefix}{A string. The first part of the output variable name. Usually
this will be "INV" but sometimes other choices will be used (e.g. "SBJ" or "IDX")}
}
\value{
A dataframe with \code{var_name} processed and renamed according to CAM standards.
}
\description{
Take in a dataframe and process a single variable to CAM standards.
}
