% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP20210729-pull_git_version.R
\name{pull_git_version}
\alias{pull_git_version}
\title{Determine Git Version for Project}
\usage{
pull_git_version()
}
\value{
A character string giving the version number, in the
  format of MAJOR.MINOR.PATCH.
}
\description{
A function to extract the git version for an R project
that has auto-generated change logs.
}
\author{
Willam Schmitt
}
