% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{pums_prep_sample_data}
\alias{pums_prep_sample_data}
\title{Process sample data to make Census comparison table}
\usage{
pums_prep_sample_data(
  demographics_target_df,
  race_var = SBJ.FCT.Race,
  sex_var = SBJ.FCT.Sex,
  hispanic_var = SBJ.FCT.Ethnicity,
  age_var = SBJ.INT.AgeAtEnrollment,
  race_map = list(White = "White", Black = "Black", Asian = "Asian",
    `Hawaiian/Pacific Islander` = "Hawaiian/Pacific Islander",
    `American Indian/Alaska Native` = "American Indian/Alaska Native", Other =
    "Middle Eastern/North African", `More than one race` = "More than one race")
)
}
\arguments{
\item{demographics_target_df}{A dataframe. The pipeline-processed demographics
for your sample}

\item{race_var}{The race variable in your target dataframe.
Defaults to \code{SBJ.FCT.Race}}

\item{sex_var}{The sex (at birth) variable in your target dataframe.
Defaults to \code{SBJ.FCT.Sex}}

\item{hispanic_var}{The ethnicity (Hispanic-indicating) variable in your
target dataframe. Defaults to \code{SBJ.FCT.Ethnicity}}

\item{race_map}{A list mapping the PUMS race values to your sample. The default
contains the mapping for the standard pipeline. The names
of the default list should be maintained (they reflect the
options on the 2023 census).}
}
\description{
\code{pums_prep_sample_data} processes sample data (usually from a pipeline-processed
demographics target dataframe) to prepare it for
making a comparison table with PUMS data.
}
