% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R19-Data_checking.R
\name{query_hist}
\alias{query_hist}
\title{Generate histogram of data query response times.}
\usage{
query_hist(df_logging, out = NULL)
}
\arguments{
\item{df_logging}{A dataframe containing processed REDCap logging information.}

\item{out}{An optional file path to save out an SVG file containing the plot.}
}
\value{
A ggplot object plotting the response times to data queries in REDCap.
}
\description{
Function that plots response times to data queries used in the REDCap data checking workflow. Expects dataframe output from \code{process_logging}.
}
