% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Data_processing_tools.R
\name{redcap_read}
\alias{redcap_read}
\title{Read All Records From a REDCap Project}
\usage{
redcap_read(redcap_uri, token, raw_or_label = "label", batch_size = 100)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the
REDCap project. Required.}

\item{token}{The user-specific string that serves to authenticate
the call to the REST API. Required.}

\item{raw_or_label}{A string (either \code{'raw'} or \code{'label'}) which
specifies whether the exported data should be the raw
coded/numerical values or the human-readable labels.
Default is label.}

\item{batch_size}{An integer indicating the number of records to
download at once.}
}
\value{
A list which contains the following elements:
\itemize{
\item \code{data}: A data frame containing all records and all fields.
\item \code{success}:A logical indicating the success of the download.
\item \code{metaData}: A data frame containing the REDCap metadata for all fields.
}
}
\description{
This function seeks to mimic the \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}} function,
but with a more robust output that can handle bizarre characters
and calculated fields. This function handles all REST API calls
using httr. This function also uses calls to \link{furrr} and \link{future} for
downloading batches from REDCap. The \link[future:plan]{plan} is not specified.
Instead, the user should use the plan that works best for their
specific use.
}
\author{
William Schmitt
}
