% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04-REDCap.R
\name{rename_redcap_vars}
\alias{rename_redcap_vars}
\title{Rename a REDCap Data Frame to Follow CAM Naming Standards}
\usage{
rename_redcap_vars(rcDtf, metaDtf)
}
\arguments{
\item{rcDtf}{A data frame containing the data downloaded from REDCap.}

\item{metaDtf}{A data frame containing the meta data (data dictionary)
downloaded from REDCap.}
}
\value{
A data frame where all columns have been renamed to conform to the
CAM standard for R names.
}
\description{
Rename a data frame downloaded from REDCap using the field annotation located
in the metadata data frame downloaded with the data. See \code{\link[=valiate_var_name]{valiate_var_name()}}
for more information on the naming standard. Warns the user for any field
that does not contain a new name that conforms to the standard in the field
annotation.
}
\examples{

# TODO

}
\author{
William Schmitt
}
