% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{run_processing_scripts}
\alias{run_processing_scripts}
\title{Run Data Processing Scripts}
\usage{
run_processing_scripts(processing_files, required_scripts, reload, debug_code)
}
\arguments{
\item{processing_files}{A vector of either...
\itemize{
\item Numeric indices denoting which processing
files to include;
\item A character vector in the form of 'DPXX'
indicating which files to include;
\item A character vector with the full file names
for the files to include.
}}

\item{required_scripts}{A vector of (complete) file names
giving the processing scripts that must be included.}

\item{reload}{Logical; if \code{TRUE}, re-downloads data
from REDCap.}

\item{debug_code}{Logical; if \code{TRUE}, displays progess at
specified cut-points for each processing script.}
}
\description{
A convenience function that navigates to the folder
with the set of data processing R scripts and
sources the specified subset of files. The portable
component of a \code{create_cleaned_data} function
typical for CAM R projects.
}
\author{
Kevin Potter
}
