% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S06-Legacy_functions.R, R/S14-save_data_files.R
\name{save_data_files}
\alias{save_data_files}
\title{Save Data Files}
\usage{
save_data_files(
  data_frames_to_save,
  extensions,
  filenames,
  filename_modifiers = NULL,
  tag = NULL,
  number = NULL,
  use_standard_names = TRUE
)

save_data_files(
  data_frames_to_save,
  extensions,
  filenames,
  filename_modifiers = NULL,
  tag = NULL,
  number = NULL,
  use_standard_names = TRUE
)
}
\arguments{
\item{data_frames_to_save}{A character vector with the names
of the data frames in the global environment that should
be saved}

\item{extensions}{A character vector with the file extensions
to output. Currently supports 'RData' and 'csv'.}

\item{filenames}{The filename(s) for the output, either a
single character string or a vector matching in length
to \code{extension}.}

\item{filename_modifiers}{A character vector of modifiers to the
filenames When saving multiple data frames as .csv files.}

\item{tag}{An optional argument for the \code{create_standardized_filename}
controlling the beginning tag at the start of the filename.}

\item{number}{An optional argument for the \code{create_standardized_filename}
specifying the file number.}

\item{use_standard_names}{Logical; if \code{TRUE} uses
the \code{create_standardized_filename} to generate filenames in
a standardized format.}
}
\description{
A convenience function that will save a
set of data frames in the global environment
as external files using the specified extension
and filenames.

A convenience function that will save a
set of data frames in the global environment
as external files using the specified extension
and filenames.
}
\author{
Kevin Potter

Kevin Potter
}
