% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R22-yaml_codebook_maker.R
\name{search_yaml_for_var}
\alias{search_yaml_for_var}
\title{Find the yaml file that contains target variable}
\usage{
search_yaml_for_var(
  varname,
  yaml_dir,
  priority_yamls = c("df_sl_subject_ids.yaml", "df_vl_visits.yaml")
)
}
\arguments{
\item{varname}{String. Name of target variable.}

\item{yaml_dir}{String. Directory containing yaml codebooks to search}

\item{priority_yamls}{Character vector. Ordered vector of yaml files to
prefer in cases where the variable is found in multiple files. The first
element of this vector has highest priority. Defaults to
\code{c(df_sl_subject_ids.yaml, df_vl_visits.yaml)}}
}
\value{
String. The name of the yaml file containing info on \code{varname}.
}
\description{
Search a yaml codebook file for a target variable
}
