% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Data_processing_tools.R
\name{source_scripts}
\alias{source_scripts}
\title{Source in Multiple Scripts in a Folder}
\usage{
source_scripts(files_to_include = NULL, path = "R")
}
\arguments{
\item{files_to_include}{A vector of either...
\itemize{
  \item Numeric indices denoting which files
    to include;
  \item A character vector with the full file names
    for the files to include.
}}

\item{path}{The folder name with the scripts to source.}
}
\description{
A convenience function that loops through
and sources files stored in a folder
located in the current working directory
(typically 'R').
}
\author{
Kevin Potter
}
