% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{subset.dictionary_meta_data}
\alias{subset.dictionary_meta_data}
\title{Extract Components From Dictionary Meta-data}
\usage{
\method{subset}{dictionary_meta_data}(x, type = "category")
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
