% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{subset.meta_data_for_variable}
\alias{subset.meta_data_for_variable}
\title{Title}
\usage{
\method{subset}{meta_data_for_variable}(x, type = "category")
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
