% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{summary.dictionary_meta_data}
\alias{summary.dictionary_meta_data}
\title{Extract Statistical Summaries From Dictionary Meta-data}
\usage{
\method{summary}{dictionary_meta_data}(x)
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
