% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S11-univariate_stat.R
\name{univariate_stat}
\alias{univariate_stat}
\title{Compute Univariate Statistic}
\usage{
univariate_stat(
  x,
  f = length,
  include = NULL,
  exclude = NULL,
  na.rm = T,
  default = NA,
  ...
)
}
\arguments{
\item{x}{A vector.}

\item{f}{A function that computes a univariate statistc
(e.g., \code{mean}, \code{length}).}

\item{include}{Optional logical vector of matching length
to \code{x} indicating obserations to include when computing
the statistic.}

\item{exclude}{Optional vector of types of observations in
\code{x} to exclude (exact matches).}

\item{na.rm}{Logical; if \code{TRUE}, removes \code{NA} values.}

\item{default}{The default value to return in the case of
missing data (i.e., no observation over which to compute a
statistic).}

\item{...}{Additional parameters for the function \code{f}.}
}
\value{
A univariate statistic, or the default value.
}
\description{
A function for flexible and robust computation of univariate
statistics over a vector of observations. Can be combined
with the \code{\link[dplyr]{group_by}} and
\code{\link[dplyr]{summarise}} functions from
the package 'dplyr'.
}
\examples{
# Load in base R data set
data( iris )

# Sepal length
sln <- iris$Sepal.Length
# Number of observations for sepal length
univariate_stat( sln )

# Mean sepal length
univariate_stat( sln, f = mean )
# Define custom function for formatted mean
f_x <- function(x) as.character( round( mean(x), 2 ) )
univariate_stat( sln, f = f_x )

# Petal length
pln <- iris$Petal.Length
# Conditional mean for sepal length when petal length < 3.5
univariate_stat( sln, f = f_x, include = pln < 3.5 )

# Species of iris
spc <- iris$Species
# Isolate species 'virginica'
vrg <- spc == 'virginica'
# No petal lengths less than 3.5 for virginica, so return default
univariate_stat( sln[vrg], f = f_x, include = pln[vrg] < 3.5, default = 'No obs' )

# Compute percentage of species 'setosa'
f_p <- function(x) paste0( round( 100*mean(x == 'setosa') ), '\%' )
univariate_stat( spc, f = f_p )
# Exclude 'virginica'
univariate_stat( spc, f = f_p, exclude = 'virginica' )

}
\author{
Kevin Potter
}
