% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{update_dictionary_meta_data}
\alias{update_dictionary_meta_data}
\title{Update Dictionary Meta-Data}
\usage{
update_dictionary_meta_data(dtf)
}
\arguments{
\item{dtf}{A data frame.}
}
\value{
A data frame whose
column attributes with dictionary
meta-data have been updated.
}
\description{
Function that updates a variable's
dictionary meta-data (specifically
the numeric summary and codes for
missing data). Useful when only
a subset of the data will be used.
}
