% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R99-Legacy.R
\name{upload_to_folder}
\alias{upload_to_folder}
\title{Upload File to Folder}
\usage{
upload_to_folder(
  filename,
  primary_path = "Dropbox",
  sub_path = NULL,
  folder_for_previous = "Previous_versions",
  filename_is_standardized = TRUE
)
}
\arguments{
\item{filename}{The file name of the file to upload
(must exist in the current working directory)}

\item{primary_path}{The primary folder where the file should be
uploaded, either...
\itemize{
\item An absolute path to the primary folder;
\item A label that can be passed to the \code{go_to} function.
}}

\item{sub_path}{An optional path to a sub-folder in the primary folder.}

\item{folder_for_previous}{An optional folder name where previous
versions of the file should be copied.}

\item{filename_is_standardized}{Logical; if \code{TRUE} the filename
is assumed to follow a standardized format, see
\code{\link[=create_standardized_filename]{create_standardized_filename()}}.}
}
\value{
A logical value, \code{TRUE} if file was successfully uploaded.
}
\description{
A convenience function that attempts to a) upload a file
to a specified folder (and optionally a sub-folder), b)
check if previous versions of the file exist and copy
them to a back-up folder.
}
\author{
Kevin Potter
}
