% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{validate_dictionary_meta_data}
\alias{validate_dictionary_meta_data}
\title{Validate Contents for Dictionary Meta-data Class Variables}
\usage{
validate_dictionary_meta_data(x)
}
\arguments{
\item{x}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
