% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{values_and_labels}
\alias{values_and_labels}
\title{Standardized Reporting for Matching Values to Labels}
\usage{
values_and_labels(values, labels)
}
\arguments{
\item{values}{A vector of codes matching in length to
\code{labels}.}

\item{labels}{A vector of labels matching in length to
\code{values}.}
}
\value{
A vector of 2 character strings.
}
\description{
Outputs character strings with standardized reporting
formats to link codes to their associated labels
(e.g., for likert-type responses) that can then be
passed into meta-data for data dictionary purposes.
}
\examples{
# Example likert-type response
v <- 1:5
l <- c( 'Never', 'Sometimes', 'Half of the time',
        'Often', 'All of the time' )
values_and_labels( v, l )

}
\author{
Kevin Potter
}
