% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions.R
\name{values_labels}
\alias{values_labels}
\title{Display Values and Associated Labels}
\usage{
values_labels(dtf, values, labels)
}
\arguments{
\item{dtf}{A data frame.}

\item{values}{A character string, the column name for
the values of interest (non-standard evaluation possible).}

\item{labels}{A character string, the column name for
the labels of interest (non-standard evaluation possible).}
}
\value{
A data frame with a column for values and
a column for associated labels.
}
\description{
A function that takes two columns in a data frame
(assumed to be an initial column of values and
a subsequent column of associated labels)
and displays the assignment of values to labels.
}
\examples{

}
\author{
Kevin Potter
}
