% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{variable_name_abbreviations}
\alias{variable_name_abbreviations}
\title{Title}
\usage{
variable_name_abbreviations(
  var_name = "",
  type = NULL,
  custom = NULL,
  n = NA,
  separator = "."
)
}
\arguments{
\item{var_name}{...}

\item{type}{...}

\item{custom}{...}

\item{n}{...}

\item{separator}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
