% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R20-Visit_and_dropout_flow.R
\name{viz_visit_flow}
\alias{viz_visit_flow}
\title{Visualize visit flow}
\usage{
viz_visit_flow(
  df_counts,
  df_visits,
  plot_title,
  output_dir = NULL,
  label_max = 5,
  show_all_visits = TRUE
)
}
\arguments{
\item{df_counts}{A dataframe with counts and record labels}

\item{df_visits}{A dataframe with participants most recent visit info.}

\item{plot_title}{A string with the title for the plot. Also used to name
svg files if \code{output_dir} is supplied.}

\item{output_dir}{(Optional) The directory where the plot should be saved}

\item{label_max}{Integer. The maximum number of record IDs that should be
printed above a visit's bar in the plot.}

\item{show_all_visits}{Logical. If TRUE, shows all visits on x-axis,
even if it was not the final visit for any participant. Defaults to TRUE.}
}
\value{
A ggplot. If \code{output_dir} is given, the plot is also saved there
as a .svg file.
}
\description{
Take in output from \code{preprocess_visit_flow_data} and plot most recent
completed visits of participants. Can be used for active participants or
dropouts.
}
