% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R18-Census_comparison.R
\name{camr_census_compare}
\alias{camr_census_compare}
\title{Process PUMS and Sample Data. Produce a table comparing the sample to the Census.}
\usage{
camr_census_compare(
  demographics_df,
  output_path,
  census_age_max = NULL,
  census_age_min = NULL,
  census_states_to_keep = NULL,
  census_states_to_drop = NULL,
  sample_data_args = NULL,
  pums_cache = "~/.cache/pums",
  census_api_key_path = NULL,
  force_pums_download = FALSE,
  age_breaks = NULL,
  age_labels = NULL,
  age_right = TRUE
)
}
\description{
Process PUMS and Sample Data. Produce a table comparing the sample to the Census.
}
