% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_ckdict}
\alias{camr_ckdict}
\title{Verify REDCap data dictionary.}
\usage{
camr_ckdict(dict)
}
\arguments{
\item{dict}{REDCap metadata table as downloaded by \code{\link[=camr_redcap_read]{camr_redcap_read()}}.}
}
\description{
Verify REDCap data dictionary.
}
\author{
Michael Pascale
}
