% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_copy_from_source}
\alias{camr_copy_from_source}
\title{Copy Files From Source Folder}
\usage{
camr_copy_from_source(
  chr_source_path = "",
  chr_destination_path = "",
  chr_source_subfolder = "",
  chr_environment = "FOLDER_SOURCE",
  chr_config = "input-root-default"
)
}
\arguments{
\item{chr_source_path}{A character string,
the absolute path to the source folder.}

\item{chr_destination_path}{A character string,
the path to the folder to which files should
be copied - if blank, uses the current working
directory.}

\item{chr_source_subfolder}{An optional character
string, the full or partial name of a subfolder
in source location with the files to copy.}

\item{chr_environment}{An optional character
string, the environmental variable with
the path to the source folder.}

\item{chr_config}{An optional character string,
the input for \code{\link[config:get]{config::get()}} to extract the
path to the source folder from a config.yml file.}
}
\value{
As a side effect copies files to a
to the specified destination folder.
}
\description{
Function to copy files from a subfolder
in a source folder to a new subfolder in
a user-defined source folder in the current
directory.
}
