% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-REDCap.R
\name{camr_download_redcap}
\alias{camr_download_redcap}
\alias{camr_redcap_download}
\title{Download Data From a REDCap Project}
\usage{
camr_download_redcap(chr_rc_uri = "", chr_rc_token = "", lgl_raw = TRUE)

camr_redcap_download(...)
}
\arguments{
\item{chr_rc_uri}{A character string, the URL
to use for the \code{httr::POST} function.}

\item{chr_rc_token}{A character string, the
user's API token for the specified REDCap
project.}

\item{lgl_raw}{Optional. Download data in raw format.
Defaults to TRUE.}
}
\value{
A list consisting of...
\itemize{
\item{'description'} {String. Explains the project from
which the data was downloaded}.
\item{'database'}{List. Information regarding the REDCap
project.}
\item{ 'data' }{Dataframe. All REDCap data with one
datapoint per row.}
\item{ 'metadata' }{Dataframe. REDCap data dictionary.}
\item{ 'form_event_map' }{Dataframe. Mappings between
REDCap events and instruments administered at each.
This will be returned for longitudinal databases only.}
}
}
\description{
Function to download data from a REDCap
Project and create a list of outputs.
}
