% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files_and_paths.R
\name{camr_filename}
\alias{camr_filename}
\title{Generates a Standard File Name}
\usage{
camr_filename(
  description,
  extension = NULL,
  project = config::get("project"),
  date = format(Sys.time(), "\%Y_\%m_\%d", tz = "UTC"),
  time = format(Sys.time(), "\%H_\%M_\%S", tz = "UTC"),
  git = TRUE
)
}
\arguments{
\item{description}{A character string. Preferably UpperCamelCase
with no spaces.}

\item{extension}{An optional character string. The file extension(s)
with no leading \code{.}.}

\item{project}{An optional character string. Name of the project.
Defaults to that stored in config.yml if present.}

\item{date}{An optional character string for the date. Preferably in
YYYY_MM_DD format - defaults to the current date.}

\item{time}{An optional character string for the time. Preferably in
HH_MM_SS format - defaults to the current time in UTC.}

\item{git}{Optional logical. The project is a git repository. Defaults
to \code{TRUE}.}
}
\value{
String. The generated file name.
}
\description{
Generates a Standard File Name
}
\author{
Michael Pascale
}
