% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R17-Table1.R
\name{camr_make_table1}
\alias{camr_make_table1}
\title{Create a table1 style table for weekly updates and
data inspection}
\usage{
camr_make_table1(
  df,
  var_label_list,
  hist_bar_color = "blue",
  group_var = NULL,
  output = NULL,
  hide_distributions = FALSE
)
}
\arguments{
\item{df}{A dataframe with columns that follow CAM-standard naming
conventions}

\item{var_label_list}{A list with names that correspond to the
variables of \code{df} that you want to summarize. The values are the
labels that will appear in the table. E.g. \code{list(SBJ.INT.Age = "Age")}}

\item{hist_bar_color}{Optional. A character or hex code for the color you'd
like the inline histograms (and horizontal bars for factor variables) to be.
Defaults to "blue".}

\item{group_var}{Optional String. The name of a factor variable by which to group
the summaries (e.g. SBJ.FCT.RandGroup to separate results by Treatment group)}

\item{output}{Optional Character. A character vector with paths as elements. Paths
can be relative to your project. If provided, the table will be saved.
Supports .html and .docx paths.}

\item{hide_distributions}{If TRUE, hide the inline distribution visualizations.
Defaults to \code{FALSE}.}
}
\value{
A gt_tbl object. Side effect: if \code{output} was provided, the table
will be saved as .html or .docx at the designated path(s)
}
\description{
This function creates a descriptive table from the specified columns
in a given dataframe. It relies on CAM variable naming conventions
to identify numeric and categorical variables. For numeric variables,
it creates an inline histogram. Optionally, you can save the table as
docx or html (note that the histograms won't render in docx)
}
\examples{
df <- data.frame(SBJ.FCT.Race = sample(c("White", "Black", "Crimean"), 100, replace=TRUE, prob = c(.5, .4, .1)),
                 SBJ.FCT.Sex = factor(sample(c("Male", "Female"), 100, replace = TRUE),
                                      levels = c("Male", "Female")),
                 SBJ.FCT.Ethnicity = factor(sample(c("Hispanic or Latino", "Not Hispanic or Latino"), 100, replace = TRUE),
                                            levels = c("Hispanic or Latino", "Not Hispanic or Latino")),
                 SBJ.INT.Age = sample(13:50, 100, replace = TRUE),
                 SBJ.FCT.Treatment = factor(sample(1:2, 100, replace = TRUE), levels = 1:2, labels = c("Group A", "Group B")))
# Factors will appear ordered by their levels
df$SBJ.FCT.Race <- df$SBJ.FCT.Race |> factor(levels = c("White", "Black", "Crimean"))
demos_to_plot <- list("SBJ.INT.Age" = "Age",
                      "SBJ.FCT.Race" = "Race",
                      "SBJ.FCT.Sex" = "Sex",
                      "SBJ.FCT.Ethnicity" = "Ethnicity")
camr_make_table1(df, demos_to_plot, group_var = "SBJ.FCT.Treatment")#, Can save out tables using arg below
            #output = c("outputs/tables/demo_table.html",
                       #"outputs/tables/demo_table.docx"))


}
\author{
Zach Himmelsbach
}
