% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camr-globals.R
\name{camr_prefix}
\alias{camr_prefix}
\title{Set a directory prefix for data files.}
\usage{
camr_prefix(prefix = getwd())
}
\description{
Set a directory prefix for data files.
}
\author{
Michael Pascale
}
