% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camrprojects-package.R
\docType{package}
\name{camrprojects-package}
\alias{camrprojects-package}
\alias{camrprojects}
\title{Portable Functions for R Projects at CAM-MGH}
\description{
The \pkg{camrprojects} package implements a set of
portable functions commonly used in R projects
from the Center for Addiction Medicine (CAM) at the
Massachusetts General Hospital (MGH). A complete list
of the functions available in the package can be
obtained via the command
\code{ls(pos = "package:camrprojects")}.
}
