% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Wrangling.R, R/S99-Legacy.R
\name{create_standardized_filename}
\alias{create_standardized_filename}
\title{Create Standardized File Name}
\usage{
create_standardized_filename(
  description,
  extension,
  tag = NULL,
  number = NULL,
  file_date = NULL,
  project_version = NULL
)

create_standardized_filename(
  description,
  extension,
  tag = NULL,
  number = NULL,
  file_date = NULL,
  project_version = NULL
)
}
\arguments{
\item{description}{A human-readable brief description of
the file content.}

\item{extension}{The file extension (e.g., 'docx', 'RData').}

\item{tag}{A single upper-case letter for file organization.}

\item{number}{A character string with a two-digit number
for file organization.}

\item{file_date}{A character string with the date
in the format 'mm_dd_yyyy'.}

\item{project_version}{The git version for the project
in the format 'v.X.X.X'.}

\item{additional}{Additional text to include following
the date. If provided, is preceded by a '-'.}

\item{date_format}{The format to use for the current
date, defaults to 'MM_DD_YYYY'.}

\item{exclude}{A vector of file names to exclude when
automatically determining file numbers}

\item{remove}{Logical; if TRUE, attempts to locate
previous versions of the outputted file name
(i.e., same name but earlier dates) and remove
them from the current folder.}
}
\value{
A character string.

A character string.
}
\description{
Create a standardized file name of the
form: TXX-Description-MM_DD_YYYY.ext where 'T'
is a leading tag, 'XX' is a file number,
'Description' is a human-readable
label, and 'ext' is a file extension.

Generates a standardized file name following the format
TXX-description-dd_mm_yyyy-v.X.X.X.extension, giving
a file tag and counter (e.g., 'F01', 'D02'), a
human-readable description and date for version control,
a machine-readable version number, and the file
extension.
}
\examples{
# Different file types
make_file_name("Example", "RData")
make_file_name("Example", "pdf")
make_file_name("Example", "docx")

# User-specified tags and numbers
make_file_name("Example", "RData", tag = "R", number = "02")
# Additional text
make_file_name("Example", "RData", additional = "v.1.0.0")
# Word document
print( create_standardized_filename(
  'Results', 'docx', project_version = 'v.1.0.0'
))
# Image file
print( create_standardized_filename(
  'Figure', 'png', project_version = 'v.1.0.0'
))

}
\author{
Kevin Potter

Kevin Potter
}
