% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S99-Legacy.R
\name{file_present}
\alias{file_present}
\title{Checks if a File is Present in a Folder}
\usage{
file_present(string, output = "Logical", std_name = FALSE)
}
\arguments{
\item{string}{A file name or part of a file name to search for.}

\item{output}{Character string indicating the type of output to
  return, either...
\enumerate{
  \item 'Logical' or 'logical'
  \item 'Vector', 'vector', or 'vec'
  \item 'Index' or 'index'
  \item 'Name' or 'name'
}}

\item{std_name}{A logical indicating whether the file follows the
standardized naming convention. If true, just matches the tag
and description of the file.}
}
\value{
Either...
\enumerate{
  \item A logical value, \code{TRUE} if the file is present;
  \item A logical vector for all files in the folder;
  \item The index position for the file if it exists;
  \item The file name.
}
}
\description{
Checks if a file is present in the current working directory.
Can check either regular file or file using MGH-CAM's standardized
file naming template: TXX-Description-MM_DD_YYYY-vX.X.X.ext
}
\author{
Kevin Potter
}
