% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{new_dictionary_meta_data}
\alias{new_dictionary_meta_data}
\title{Constructor Function for Dictionary Meta-Data Class}
\usage{
new_dictionary_meta_data(
  Column_name = "",
  Variable_category = "",
  Data_type = "",
  Sub_category = "",
  Description = "",
  REDCap_variables = "",
  Values_and_labels = "",
  Scale = "",
  Subscale = "",
  Summary = "",
  Units = "",
  Codes_for_missing = "",
  Groups_collected_over = "",
  Times_collected_over = "",
  Studies_collected_over = "",
  Validated = "",
  Notes = ""
)
}
\arguments{
\item{Column_name}{A standardized column name,
(e.g., 'GGG.TTT.Var_name' or
'GGG.TTT.Optional.Var_name', where 'GGG' is
an abbreviation for the overarching group for a
variable, 'TTT' is an abbreviation for the data type,
and 'Optional' can (if specified) refer to
abbreviations for a sub-category).}

\item{Variable_category}{The label associated with
the 3-letter abbreviation giving the overarching
variable group to which the column belongs.}

\item{Data_type}{The label associated with the
3-letter abbreviation for the data type of
the column.}

\item{Sub_category}{If specified, the label associated
with the optional abbreviation for any sub-categories.}

\item{Description}{A brief description of what the
variable is.}

\item{REDCap_variables}{If specified, a character vector
with the variable names from the raw data frame downloaded
from REDCap used to create the current column.}

\item{Values_and_labels}{If specified, a named list
with two vectors, \code{Values} and \code{Labels},
giving the raw values and the corresponding
labels (e.g., for Likert-scale responses).}

\item{Scale}{If specified, a list ..}

\item{Subscale}{If specified, a list...}

\item{Summary}{If specified, a list with a descriptive
statistical summary of the non-missing values for
the column.}

\item{Units}{If specified, a character string describing
the units of measurement for the column.}

\item{Codes_for_missing}{A list of the codes (e.g.,
\code{NA}, \code{''}, etc.) for missing values.}

\item{Groups_collected_over}{A vector of values
describing the study groups for which the
current variable was collected.}

\item{Times_collected_over}{A vector of values
describing the study time points during which the
current variable was collected.}

\item{Studies_collected_over}{A vector of values
describing the studies (e.g., a pilot study,
the active study), for which the current variable
was collected.}

\item{Validated}{A character string briefly describing
if and how the values in the column were validated
and data checked.}

\item{Notes}{A character string with any additional
notes.}
}
\value{
A named list of class \code{dictionary_meta_data}.
}
\description{
A function to create a list with the proper
slots for the class \code{dictionary_meta_data}.
}
\author{
Kevin Potter
}
