% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S02-Data_processing_tools.R
\name{postFormJSONToDf}
\alias{postFormJSONToDf}
\title{Sends a POST Form via httr and Converts JSON Return to a Data Frame}
\usage{
postFormJSONToDf(url, params)
}
\arguments{
\item{url}{The URL where the POST request should be submitted.}

\item{params}{A named list of parameters to submit with the POST
request.}
}
\value{
A list which contains the following elements:
\itemize{
\item \code{data}: A data frame containing all records and all fields.
\item \code{status}: The HTTP status code returned from the request. If an
error was encountered after the request, returns -1.
}
}
\description{
This function uses httr to submit a POST form to the given url and
converts the response (expected to be a JSON array) to a data.frame.
Returns a list containing the data frame and response from the server.
}
\author{
William Schmitt
}
