% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{print.dictionary_meta_data}
\alias{print.dictionary_meta_data}
\title{Display Dictionary Meta-data Information}
\usage{
\method{print}{dictionary_meta_data}(x, digits = 2)
}
\arguments{
\item{x}{An R object of class \code{dictionary_meta_data}.}
}
\value{
The method prints to console, at a minimum, the
the variable's overarching group, data type, and
associated description. Additional information may include
the original REDCap variable used to create the current
variable, the unit of measurement, a brief statistical
summary when appropriate, and details for scales,
inventory, or questionnaire measures.
}
\description{
A print method for the \code{dictionary_meta_data}
class, printing to the console key information
contained within the dictionary meta-data fields.
}
\examples{
# Examples

}
