% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S20-redcap_read.R
\name{redcap_read}
\alias{redcap_read}
\title{Read all records from a REDCap project}
\usage{
redcap_read(redcap_uri, token, raw_or_label = "label", batch_size = 100)
}
\arguments{
\item{redcap_uri}{The URI (uniform resource identifier) of the
REDCap project. Required.}

\item{token}{The user-specific string that serves to authenticate
the call to the REST API. Required.}

\item{raw_or_label}{A string (either `'raw'` or `'label'`) which
specifies whether the exported data should be the raw
coded/numerical values or the human-readable labels.
Default is label.}

\item{batch_size}{An integer indicating the number of records to
download at once.}
}
\value{
A list which contains the following elements:
* `data`: A data frame containing all records and all fields.
* `success`:A logical indicating the success of the download.
}
\description{
This function seeks to mimic the [REDCapR::redcap_read()] function,
but with a more robust output that can handle bizarre characters
and calculated fields. Instead of relying on base R HTTP protocols,
this function handles all REST API calls using RCurl.
}
\author{
William Schmitt
}
