% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{scale_format}
\alias{scale_format}
\title{Standardized Reporting Format for Scales/Inventories/Questionnaires}
\usage{
scale_format(
  name,
  n_items,
  range,
  abbreviation = "",
  cut_off = NA,
  reference = "",
  subscale = FALSE
)
}
\arguments{
\item{name}{A character string with the full name of
the scale (e.g., 'Hospital Anxiety Depression Scale'
or the 'Cannabis Use Disorder Identification Test').}

\item{n_items}{An integer with the number of
questions/items in the scale (or subscale).}

\item{range}{...}

\item{abbreviation}{A character string with the abbreviation
for the scale (e.g., 'HADS' or 'CUDIT').}

\item{cut_off}{An optional numeric vector giving any cut-offs
used in the scale (e.g., the CUDIT uses a clinical cut-off
of scores at \code{8} or higher to indicate problematic
cannabis use).}

\item{reference}{...}

\item{subscale}{Logical; if \code{TRUE}, instead outputs
the reporting format for subscales (consisting only
of the subscale name its number of items).}
}
\value{
...
}
\description{
Given a scale's name, number of items, and abbreviation,
creates a character string following a standardized
reporting format to pass into meta-data for data
dictionary purposes. Provides additional options for
reporting cut-offs (e.g., clinical cut-offs used for
diagnosis purposes) and subscale information.
}
\examples{
scale_format(
  'Cannabis Use Disorder Identification Test',
  'CUDIT',
  8
)

}
\author{
Kevin Potter
}
