% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Metadata.R
\name{scale_format}
\alias{scale_format}
\title{Standardized Reporting Format for Scales/Inventories/Questionnaires}
\usage{
scale_format(
  name,
  n_items,
  range,
  abbreviation = "",
  cut_off = NA,
  reference = "",
  interpretation = "",
  description = "",
  units_of_x = "",
  subscale = FALSE
)
}
\arguments{
\item{name}{A character string with the full name of
the scale (e.g., 'Hospital Anxiety Depression Scale'
or the 'Cannabis Use Disorder Identification Test').}

\item{n_items}{An integer with the number of
questions/items in the scale (or subscale).}

\item{range}{A vector with the lower and upper limits
for the possible scores the measure can have.}

\item{abbreviation}{A character string with the abbreviation
for the scale (e.g., 'HADS' or 'CUDIT').}

\item{cut_off}{An optional numeric vector giving any cut-offs
used in the scale (e.g., the CUDIT uses a clinical cut-off
of scores at \code{8} or higher to indicate problematic
cannabis use).}

\item{reference}{A character vector with APA-style
references that describe the measure and relevant
details on its scoring.}

\item{interpretation}{A brief note on how to interpret
measure scores (e.g., describe what higher scores
indicate, such as how higher scores on the CUDIT
indicate more problematic cannabis use).}

\item{subscale}{Logical; if \code{TRUE}, instead outputs
the reporting format for subscales (consisting only
of the subscale name its number of items).}
}
\value{
A list.
}
\description{
A function to help structure data dictionary information for
scales/inventories/questionnaires into a standardized
format.
}
\examples{
scale_format(
  name = 'Cannabis Use Disorder Identification Test (Revised)',
  n_items = 8,
  range = c( 0, 32 ),
  abbreviation = 'CUDIT-R',
  interpretation = 'Higher scores indicate more problematic cannabis use'
)

}
\author{
Kevin Potter
}
