% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S05-Meta_data_functions.R
\name{validate_dictionary_meta_data}
\alias{validate_dictionary_meta_data}
\title{Validate Contents for Dictionary Meta-data Class Variables}
\usage{
validate_dictionary_meta_data(x)
}
\arguments{
\item{x}{An R object of class \code{dictionary_meta_data}.}
}
\value{
If any content is misspecified for \code{x}
returns an informative error message.
}
\description{
A function to check whether the content for
R objects of class \code{dictionary_meta_data}
have been correctly specified.
}
